Module Preface

The library is divided into 3 parts (in the user area) which serve complementary purposes.

Abstraction implementations

Functor (in Haskell sense), Applicatives and monads are some of the best known abstractions in functional programming. Indeed, they allow recurrent problems to be solved in an elegant way. Generally, thanks to certain mechanisms linked to the languages that implement them (in Haskell, for example, using typeclasses), it is possible, by defining only a small subset of their combinators, to derive many others. So the purpose of "this part of the library" is to provide mechanisms for deriving combinators for a given type and a chosen abstraction, respecting OCaml programming idioms as much as possible.

Specifications

This module describes the specifications of the abstractions provided by Preface. These specifications, which correspond to interfaces (module types in OCaml terminology) serve as constraints for the functors described in Preface.Make and centralise the documentation. Using a separate module allows cyclic dependencies to be resolved if one module can be described by another module and vice versa.

module Specs = Preface_specs

Concretisations

In order to produce embodiments for the abstractions described in Preface.Specs, Preface.Make offers a collection of functors that take modules constrained by the interfaces described in Preface.Specs to produce modules that respect the more complete interfaces also described in Preface.Specs.

module Make = Preface_make

Concepts, Naming and Terminology

The modular design of Preface may seem a little intimidating at first glance. Let's look at the logic of the cut to understand how best to use it to describe new concretizations of abstractions.

Abstractions must respect a minimum interface, however, sometimes there are several paths to describe the abstraction. For example, building a Monad on a type requires a return (or pure depending on the convention in practice) and:

bind/>>=

or map and join

or sometimes >=>

In addition, on the basis of these minimum combinators, it is possible to derive other combinators. However, it happens that these combinators are not implemented in an optimal way (this is the cost of abstraction). In the OCaml ecosystem, the use of polymorphic variants is sometimes used to give the user the freedom to implement, or not, a function by wrapping the function definition in a value of this type:

val f : [< `Derived | `Custom of 'a -> 'b ] 

Instead of relying on this kind of (rather clever!) trick, we decided to rely mainly on the module language.

To make it easy to describe the embodiment of an abstraction, but still allow for the possibility of providing more efficient implementations (that propagate new implementations on aliases, such as infix operators, or functions that use these functions), Preface proposes a rather particular cut.

Each abstraction is broken down into several sub-modules:

Standard library

Whereas the previous section dealt mainly with the concretisation of abstractions (using functor machinery). This section documents the standard Preface library. A collection of already implemented abstractions for relatively common data structures.

Common datatypes

module Identity = Preface_stdlib.Identity

module Option = Preface_stdlib.Option

module Either = Preface_stdlib.Either

module Pair = Preface_stdlib.Pair

Collection

module List = Preface_stdlib.List

module Nonempty_list = Preface_stdlib.Nonempty_list

module Stream = Preface_stdlib.Stream

Error handling

module Exn = Preface_stdlib.Exn

module Result = Preface_stdlib.Result

module Validation = Preface_stdlib.Validation

module Try = Preface_stdlib.Try

module Validate = Preface_stdlib.Validate

Functions

module Fun = Preface_stdlib.Fun

module Predicate = Preface_stdlib.Predicate

module Continuation = Preface_stdlib.Continuation

Transformers over identity

There are some (monad or comonad) transformers defined in Spec/Make. In Stdlib these are some concretised version using Identity as inner monad or comonad.

module Reader = Preface_stdlib.Reader

module Writer = Preface_stdlib.Writer

module State = Preface_stdlib.State

module Store = Preface_stdlib.Store

module Env = Preface_stdlib.Env

module Traced = Preface_stdlib.Traced

Static Analysis

Applicatives, Selectives, Profunctors and Arrows allow, contrary to monads, to perform static analyses on calculation workflows. Over and Under allow optimistic or pessimistic approximations.

module Approximation = Preface_stdlib.Approximation