Parameter For_rigid.S

Type

type ('a, 'index) t

The type held by the Indexed Selective.

Functions

val select : (('a, 'b) Stdlib.Either.t, 'index) t -> ('a -> 'b, 'index) t -> ('b, 'index) t

select e f apply f if e is Left. It allow to skip effect using Right.

val branch : (('a, 'b) Stdlib.Either.t, 'index) t -> ('a -> 'c, 'index) t -> ('b -> 'c, 'index) t -> ('c, 'index) t

branch is like select. It chooses between two effects.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_apply.WITH_APPLY with type ('a, 'index) t := ('a, 'index) t

val apply : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

May apply a function wrapped into an t to a value also wrapped into an t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_functor.WITH_MAP with type ('a, 'index) t := ('a, 'index) t

val map : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

include Preface_specs.Indexed_apply.WITH_PRODUCT with type ('a, 'index) t := ('a, 'index) t

val product : ('a, 'index) t -> ('b, 'index) t -> ('a * 'b, 'index) t

Monoidal product between two t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := ('a, 'index) t

val pure : 'a -> ('a, 'index) t

Lift a value from 'a into a new t.

include Preface_specs.Indexed_apply.WITH_LIFT2 with type ('a, 'index) t := ('a, 'index) t

val lift2 : ('a -> 'b -> 'c) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val lift3 : ('a -> 'b -> 'c -> 'd) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t -> ('d, 'index) t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val replace : 'a -> ('b, 'index) t -> ('a, 'index) t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : ('a, 'index) t -> (unit, 'index) t

Create a new t, replacing all values in the given functor by unit.

val if_ : (bool, 'index) t -> ('a, 'index) t -> ('a, 'index) t -> ('a, 'index) t

Same of branch but using a Boolean as disjunction.

val bind_bool : (bool, 'index) t -> (bool -> ('a, 'index) t) -> ('a, 'index) t

Monad bind specialized for Boolean.

val when_ : (bool, 'index) t -> (unit, 'index) t -> (unit, 'index) t

Conditionally perform an effect.

val exists : ('a -> (bool, 'index) t) -> 'a list -> (bool, 'index) t

Selective version of List.exists.

val for_all : ('a -> (bool, 'index) t) -> 'a list -> (bool, 'index) t

Selective version of List.for_all.

val or_ : (bool, 'index) t -> (bool, 'index) t -> (bool, 'index) t

Or combinator.

val and_ : (bool, 'index) t -> (bool, 'index) t -> (bool, 'index) t

And combinator.

val while_ : (bool, 'index) t -> (unit, 'index) t

Keep checking an effectful condition while it holds.

Infix operators

module Infix : Preface_specs.Indexed_selective.INFIX with type ('a, 'index) t := ('a, 'index) t

val (<*>) : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (<**>) : ('a, 'index) t -> ('a -> 'b, 'index) t -> ('b, 'index) t

Flipped Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (*>) : (unit, 'index) t -> ('a, 'index) t -> ('a, 'index) t

Discard the value of the first argument.

val (<*) : ('a, 'index) t -> (unit, 'index) t -> ('a, 'index) t

Discard the value of the second argument.

val (<$>) : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Infix version of CORE.map.

val (<&>) : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

Flipped and infix version of CORE.map.

val (<$) : 'a -> ('b, 'index) t -> ('a, 'index) t

Infix version of OPERATION.replace.

val ($>) : ('a, 'index) t -> 'b -> ('b, 'index) t

Flipped and infix version of OPERATION.replace.

val (<*?) : (('a, 'b) Stdlib.Either.t, 'index) t -> ('a -> 'b, 'index) t -> ('b, 'index) t

Infix version of CORE.select.

val (<||>) : (bool, 'index) t -> (bool, 'index) t -> (bool, 'index) t

Infix version of CORE.or_.

val (<&&>) : (bool, 'index) t -> (bool, 'index) t -> (bool, 'index) t

Infix version of CORE.and_.

Syntax

module Syntax : Preface_specs.Indexed_selective.SYNTAX with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_apply.SYNTAX with type ('a, 'index) t := ('a, 'index) t

val let+ : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

let operator for mapping.

val and+ : ('a, 'index) t -> ('b, 'index) t -> ('a * 'b, 'index) t

and operator for the monoidal product.