Module Alt.Operation

Parameters

Core : Preface_specs.Alt.CORE

Signature

type 'a t = 'a Core.t

A type 'a t held by the Alt.

val times : int -> 'a t -> 'a t option

times n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : 'a t Preface_core.Nonempty_list.t -> 'a t

Reduce a Nonempty_list.t using combine.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.