Module Alternative.Via_lift2

Parameters

Req : Preface_specs.Alternative.WITH_LIFT2

Signature

Type

type 'a t = 'a Req.t

The type held by the Alternative.

Functions

type 'a t

A type 'a t held by the Alt.

val combine : 'a t -> 'a t -> 'a t

Combine two values of 'a t into one.

val neutral : 'a t

The neutral element of the Alternative.

type 'a t

The type held by the Applicative.

val pure : 'a -> 'a t

Lift a value from 'a into a new 'a t.

val apply : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over 'a t to 'b t.

type 'a t

The type held by the Applicative.

val pure : 'a -> 'a t

Lift a value from 'a into a new 'a t.

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val product : 'a t -> 'b t -> ('a * 'b) t

Product functor mapping from 'a t and 'b t to ('a * 'b) t.

type 'a t

The type held by the Applicative.

val pure : 'a -> 'a t

Lift a value from 'a into a new 'a t.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Mapping over from 'a and 'b to 'c over 'a t and 'b t to 'c t.

type 'a t

The type held by the Applicative.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Mapping over from 'a and 'b and 'c to 'd over 'a t and 'b t and 'c t to 'd t.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

type 'a t

A type 'a t held by the Alt.

val times : int -> 'a t -> 'a t option

times n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : 'a t Preface_core.Nonempty_list.t -> 'a t

Reduce a Nonempty_list.t using combine.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

val reduce : 'a t list -> 'a t

Reduce a List.t using combine.

Infix operators

module Infix : Preface_specs.Alternative.INFIX with type 'a t := 'a t

type 'a t

The type held by the Applicative.

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over 'a t to 'b t.

val (<**>) : 'a t -> ('a -> 'b) t -> 'b t

Flipped Applicative functor of ('a -> 'b) t over 'a t to 'b t.

val (*>) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (<*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

type 'a t

The type held by the Functor.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Preface_specs.Functor.CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Preface_specs.Functor.CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Preface_specs.Functor.OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Preface_specs.Functor.OPERATION.replace.

type 'a t

A type 'a t which is an Alt.

val (<|>) : 'a t -> 'a t -> 'a t

Infix version of Alt.CORE.combine

type 'a t

The type held by the Functor.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Preface_specs.Functor.CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Preface_specs.Functor.CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Preface_specs.Functor.OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Preface_specs.Functor.OPERATION.replace.

Syntax

module Syntax : Preface_specs.Alternative.SYNTAX with type 'a t := 'a t

type 'a t

The type held by the Applicative.

val let+ : 'a t -> ('a -> 'b) -> 'b t

Flipped mapping over from 'a to 'b over 'a t to 'b t.

val and+ : 'a t -> 'b t -> ('a * 'b) t

Product functor mapping from 'a t and 'b t to ('a * 'b) t.