Module Preface_make.Applicative

Building a Preface_specs.Applicative

Using the minimal definition

Using pure and apply

Build a Preface_specs.APPLICATIVE using Preface_specs.Applicative.WITH_APPLY.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_apply : functor (Req : Preface_specs.Applicative.WITH_APPLY) -> Preface_specs.APPLICATIVE with type 'a t = 'a Req.t

Using pure, map and product

Build a Preface_specs.APPLICATIVE using Preface_specs.Applicative.WITH_MAP_AND_PRODUCT.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_product : functor (Req : Preface_specs.Applicative.WITH_MAP_AND_PRODUCT) -> Preface_specs.APPLICATIVE with type 'a t = 'a Req.t

Using pure and lift2

Build a Preface_specs.APPLICATIVE using Preface_specs.Applicative.WITH_LIFT2.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_lift2 : functor (Req : Preface_specs.Applicative.WITH_LIFT2) -> Preface_specs.APPLICATIVE with type 'a t = 'a Req.t

Applicative Algebra

Construction of Preface_specs.APPLICATIVE by combining them.

Composition

Right-to-left composition of Preface_specs.APPLICATIVE.

module Composition : functor (F : Preface_specs.APPLICATIVE) -> functor (G : Preface_specs.APPLICATIVE) -> Preface_specs.APPLICATIVE with type 'a t = 'a G.t F.t

Product

Construct the product of two Preface_specs.APPLICATIVE.

module Product : functor (F : Preface_specs.APPLICATIVE) -> functor (G : Preface_specs.APPLICATIVE) -> Preface_specs.APPLICATIVE with type 'a t = 'a F.t * 'a G.t

From other abstraction

From a Monad

Produces a Preface_specs.APPLICATIVE from a Preface_specs.MONAD.

module From_monad : functor (Monad : Preface_specs.MONAD) -> Preface_specs.APPLICATIVE with type 'a t = 'a Monad.t

From an Alternative

Produces a Preface_specs.APPLICATIVE from an Preface_specs.ALTERNATIVE.

module From_alternative : functor (Alternative : Preface_specs.ALTERNATIVE) -> Preface_specs.APPLICATIVE with type 'a t = 'a Alternative.t

From an Arrow

Produces a Preface_specs.APPLICATIVE from an Preface_specs.ARROW.

module From_arrow : functor (A : Preface_specs.ARROW) -> Preface_specs.APPLICATIVE with type 'a t = (unit, 'a) A.t

From a Monoid

Produces a Preface_specs.APPLICATIVE from a Preface_specs.MONOID. This Applicative is called Const or a phantom monoid.

module Const : functor (M : Preface_specs.Monoid.CORE) -> sig ... end

Manual construction

Advanced way to build an Preface_specs.APPLICATIVE, constructing and assembling a component-by-component of Preface_specs.APPLICATIVE. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Applicative.CORE) -> functor (Operation : Preface_specs.Applicative.OPERATION with type 'a t = 'a Core.t) -> functor (Infix : Preface_specs.Applicative.INFIX with type 'a t = 'a Core.t) -> functor (Syntax : Preface_specs.Applicative.SYNTAX with type 'a t = 'a Core.t) -> Preface_specs.APPLICATIVE with type 'a t = 'a Core.t

Building Core

module Core_via_map_and_product : functor (Req : Preface_specs.Applicative.WITH_MAP_AND_PRODUCT) -> Preface_specs.Applicative.CORE with type 'a t = 'a Req.t

module Core_via_apply : functor (Req : Preface_specs.Applicative.WITH_APPLY) -> Preface_specs.Applicative.CORE with type 'a t = 'a Req.t

module Core_via_lift2 : functor (Req : Preface_specs.Applicative.WITH_LIFT2) -> Preface_specs.Applicative.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Applicative.CORE) -> Preface_specs.Applicative.OPERATION with type 'a t = 'a Core.t

Deriving Syntax

module Syntax : functor (Core : Preface_specs.Applicative.CORE) -> Preface_specs.Applicative.SYNTAX with type 'a t = 'a Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Applicative.CORE) -> functor (Operation : Preface_specs.Applicative.OPERATION with type 'a t = 'a Core.t) -> Preface_specs.Applicative.INFIX with type 'a t = 'a Core.t