Module Preface_make.Apply

Building a Preface_specs.APPLY

Using the minimal definition

Using map and apply

Build a Preface_specs.APPLY using Preface_specs.APPLY.WITH_MAP_AND_APPLY.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_apply (Req : Preface_specs.Apply.WITH_MAP_AND_APPLY) : Preface_specs.APPLY with type 'a t = 'a Req.t

Using apply over functor

Build a Preface_specs.APPLY using Preface_specs.Functor.WITH_MAP and Preface_specs.Apply.WITH_APPLY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_functor_via_apply (Functor : Preface_specs.Functor.WITH_MAP) (Req : Preface_specs.Apply.WITH_APPLY with type 'a t = 'a Functor.t) : Preface_specs.APPLY with type 'a t = 'a Req.t

Using map and product

Build a Preface_specs.APPLY using Preface_specs.Apply.WITH_MAP_AND_PRODUCT.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_product (Req : Preface_specs.Apply.WITH_MAP_AND_PRODUCT) : Preface_specs.APPLY with type 'a t = 'a Req.t

Using product over functor

Build a Preface_specs.APPLY using Preface_specs.Functor.WITH_MAP and Preface_specs.Apply.WITH_PRODUCT.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_functor_via_product (Functor : Preface_specs.Functor.WITH_MAP) (Req : Preface_specs.Apply.WITH_PRODUCT with type 'a t = 'a Functor.t) : Preface_specs.APPLY with type 'a t = 'a Req.t

Using map and lift2

Build a Preface_specs.APPLY using Preface_specs.Apply.WITH_LIFT2.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_lift2 (Req : Preface_specs.Apply.WITH_MAP_AND_LIFT2) : Preface_specs.APPLY with type 'a t = 'a Req.t

Using lift2 over functor

Build a Preface_specs.APPLY using Preface_specs.Functor.WITH_MAP and Preface_specs.Apply.WITH_LIFT2.

Other standard method, using the minimal definition of an alt to derive its full API.

module Over_functor_via_lift2 (Functor : Preface_specs.Functor.WITH_MAP) (Req : Preface_specs.Apply.WITH_LIFT2 with type 'a t = 'a Functor.t) : Preface_specs.APPLY with type 'a t = 'a Req.t

APPLY Algebra

Construction of Preface_specs.APPLY by combining them.

Composition

Right-to-left composition of Preface_specs.APPLY.

module Composition (F : Preface_specs.APPLY) (G : Preface_specs.APPLY) : Preface_specs.APPLY with type 'a t = 'a G.t F.t

Product

Construct the product of two Preface_specs.APPLY.

module Product (F : Preface_specs.APPLY) (G : Preface_specs.APPLY) : Preface_specs.APPLY with type 'a t = 'a F.t * 'a G.t

From other abstraction

From a Monad

Produces a Preface_specs.APPLY from a Preface_specs.MONAD.

module From_bind (Bind : Preface_specs.Bind.CORE) : Preface_specs.APPLY with type 'a t = 'a Bind.t

From a Monad

Produces a Preface_specs.APPLY from a Preface_specs.MONAD.

module From_monad (Monad : Preface_specs.MONAD) : Preface_specs.APPLY with type 'a t = 'a Monad.t

From an Applicative

Produces a Preface_specs.APPLY from a Preface_specs.MONAD.

module From_applicative (Applicative : Preface_specs.APPLICATIVE) : Preface_specs.APPLY with type 'a t = 'a Applicative.t

From an Alternative

Produces a Preface_specs.APPLY from an Preface_specs.ALTERNATIVE.

module From_alternative (Alternative : Preface_specs.ALTERNATIVE) : Preface_specs.APPLY with type 'a t = 'a Alternative.t

From an Arrow

Produces a Preface_specs.APPLY from an Preface_specs.ARROW.

module From_arrow (A : Preface_specs.ARROW) : Preface_specs.APPLY with type 'a t = (unit, 'a) A.t

From a Monoid

Produces a Preface_specs.APPLY from a Preface_specs.MONOID. This APPLY is called Const or a phantom monoid.

module Const (M : Preface_specs.Monoid.CORE) : sig ... end

To other abstraction

To an Indexed Apply

module Index (F : Preface_specs.APPLY) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = 'a F.t

Manual construction

Advanced way to build an Preface_specs.APPLY, constructing and assembling a component-by-component of Preface_specs.APPLY. (In order to provide your own implementation for some features.)

Grouping of all components

module Via (Core : Preface_specs.Apply.CORE) (Operation : Preface_specs.Apply.OPERATION with type 'a t = 'a Core.t) (Infix : Preface_specs.Apply.INFIX with type 'a t = 'a Core.t) (Syntax : Preface_specs.Apply.SYNTAX with type 'a t = 'a Core.t) : Preface_specs.APPLY with type 'a t = 'a Core.t

Building Core

module Core_via_map_and_apply (Req : Preface_specs.Apply.WITH_MAP_AND_APPLY) : Preface_specs.Apply.CORE with type 'a t = 'a Req.t

module Core_via_map_and_product (Req : Preface_specs.Apply.WITH_MAP_AND_PRODUCT) : Preface_specs.Apply.CORE with type 'a t = 'a Req.t

module Core_via_map_and_lift2 (Req : Preface_specs.Apply.WITH_MAP_AND_LIFT2) : Preface_specs.Apply.CORE with type 'a t = 'a Req.t

module Core_over_functor_via_apply (Functor : Preface_specs.Functor.WITH_MAP) (Req : Preface_specs.Apply.WITH_APPLY with type 'a t = 'a Functor.t) : Preface_specs.Apply.CORE with type 'a t = 'a Req.t

module Core_over_functor_via_product (Functor : Preface_specs.Functor.WITH_MAP) (Req : Preface_specs.Apply.WITH_PRODUCT with type 'a t = 'a Functor.t) : Preface_specs.Apply.CORE with type 'a t = 'a Req.t

module Core_over_functor_via_lift2 (Functor : Preface_specs.Functor.WITH_MAP) (Req : Preface_specs.Apply.WITH_LIFT2 with type 'a t = 'a Functor.t) : Preface_specs.Apply.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation (Core : Preface_specs.Apply.CORE) : Preface_specs.Apply.OPERATION with type 'a t = 'a Core.t

Deriving Syntax

module Syntax (Core : Preface_specs.Apply.CORE) : Preface_specs.Apply.SYNTAX with type 'a t = 'a Core.t

Deriving Infix

module Infix (Core : Preface_specs.Apply.CORE) (Operation : Preface_specs.Apply.OPERATION with type 'a t = 'a Core.t) : Preface_specs.Apply.INFIX with type 'a t = 'a Core.t