Parameter Via.1-Core

type ('a, 'b) t

The type held by the Arrow.

type ('a, 'b) t

The type held by the Category.

val id : ('a, 'a) t

The identity morphism.

val compose : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

Morphism composition (from right to left).

type ('a, 'b) t

The type held by the Arrow.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val fst : ('a, 'b) t -> ('a * 'd, 'b * 'd) t

Send the first component of the input through the argument Arrow, and copy the rest unchanged to the output.

type ('a, 'b) t

The type held by the Arrow.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val split : ('a, 'b) t -> ('c, 'd) t -> ('a * 'c, 'b * 'd) t

Split the input between the two given Arrows and combine their output.