Parameter Core_over_category_and_via_arrow_and_fst.2-Req

type ('a, 'b) t = ('a, 'b) Category.t

The type held by the Arrow_alt.

val combine : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

Combine two values of ('a, 'b) t into one.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val fst : ('a, 'b) t -> ('a * 'd, 'b * 'd) t

Send the first component of the input through the argument Arrow, and copy the rest unchanged to the output.