Parameter Over_category_and_via_arrow_and_fst.Req

type ('a, 'b) t = ('a, 'b) Category.t

The type held by the Arrow_apply.

val apply : (('a, 'b) t * 'a, 'b) t

application of an arrow to an input.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val fst : ('a, 'b) t -> ('a * 'd, 'b * 'd) t

Send the first component of the input through the argument Arrow, and copy the rest unchanged to the output.