Module Preface_make.Arrow_choice

Building a Preface_specs.Arrow_choice

Using the minimal definition

Using left over an Arrow

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_LEFT over an Preface_specs.ARROW.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_arrow_with_left : functor (Arrow : Preface_specs.ARROW) -> functor (Left : Preface_specs.Arrow_choice.WITH_LEFT with type ('a, 'b) t = ('a, 'b) Arrow.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Left.t

Using choose over an Arrow

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_CHOOSE over an Preface_specs.ARROW.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_arrow_with_choose : functor (Arrow : Preface_specs.ARROW) -> functor (Choose : Preface_specs.Arrow_choice.WITH_CHOOSE with type ('a, 'b) t = ('a, 'b) Arrow.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Choose.t

Using left and choose over an Arrow

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_LEFT_AND_CHOOSE over an Preface_specs.ARROW.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_arrow_with_left_and_choose : functor (Arrow : Preface_specs.ARROW) -> functor (Choose_left : Preface_specs.Arrow_choice.WITH_LEFT_AND_CHOOSE with type ('a, 'b) t = ('a, 'b) Arrow.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Choose_left.t

Using left, arrow and fst over a Category

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_ARROW_AND_FST_AND_LEFT over a Preface_specs.CATEGORY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_category_and_via_arrow_and_fst_and_left : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_FST_AND_LEFT with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Req.t

Using choose, arrow and fst over a Category

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_ARROW_AND_FST_AND_CHOOSE over a Preface_specs.CATEGORY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_category_and_via_arrow_and_fst_and_choose : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_FST_AND_CHOOSE with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Req.t

Using left, arrow and split over a Category

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_ARROW_AND_SPLIT_AND_LEFT over a Preface_specs.CATEGORY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_over_category_and_via_arrow_and_split_and_left : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_SPLIT_AND_LEFT with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Req.t

Using choose, arrow and split over a Category

Build an Preface_specs.ARROW_CHOICE using Preface_specs.Arrow_choice.WITH_ARROW_AND_SPLIT_AND_CHOOSE over a Preface_specs.CATEGORY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_category_and_via_arrow_and_split_and_choose : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_SPLIT_AND_CHOOSE with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Req.t

Arrow Choice Algebra

Construction of Preface_specs.ARROW_CHOICE by combining them.

Product

Construct the product of two Preface_specs.ARROW_CHOICE.

module Product : functor (F : Preface_specs.ARROW_CHOICE) -> functor (G : Preface_specs.ARROW_CHOICE) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) F.t * ('a, 'b) G.t

From other abstraction

From a Monad

Produces an Preface_specs.ARROW_CHOICE from a Preface_specs.MONAD (using the Kleisli Arrow).

module From_monad : functor (Monad : Preface_specs.Monad.CORE) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = 'a -> 'b Monad.t

Manual construction

Advanced way to build an Preface_specs.ARROW_CHOICE, constructing and assembling a component-by-component of Preface_specs.ARROW_CHOICE. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Arrow_choice.CORE) -> functor (Operation : Preface_specs.Arrow_choice.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t) -> functor (Alias : Preface_specs.Arrow_choice.ALIAS with type ('a, 'b) t = ('a, 'b) Operation.t) -> functor (Infix : Preface_specs.Arrow_choice.INFIX with type ('a, 'b) t = ('a, 'b) Alias.t) -> Preface_specs.ARROW_CHOICE with type ('a, 'b) t = ('a, 'b) Infix.t

Building Core

module Core_over_category_and_via_arrow_and_fst_and_left : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_FST_AND_LEFT with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_choice.CORE with type ('a, 'b) t = ('a, 'b) Req.t

module Core_over_category_and_via_arrow_and_split_and_left : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_SPLIT_AND_LEFT with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_choice.CORE with type ('a, 'b) t = ('a, 'b) Req.t

module Core_over_category_and_via_arrow_and_fst_and_choose : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_FST_AND_CHOOSE with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_choice.CORE with type ('a, 'b) t = ('a, 'b) Req.t

module Core_over_category_and_via_arrow_and_split_and_choose : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_choice.WITH_ARROW_AND_SPLIT_AND_CHOOSE with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_choice.CORE with type ('a, 'b) t = ('a, 'b) Req.t

Deriving Operation

module Operation_over_category : functor (Category : Preface_specs.CATEGORY) -> functor (Core : Preface_specs.Arrow_choice.CORE with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_choice.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t

Deriving Alias

module Alias : functor (Operation : Preface_specs.Arrow.OPERATION) -> Preface_specs.Arrow.ALIAS with type ('a, 'b) t = ('a, 'b) Operation.t

Deriving Infix

module Infix_over_category : functor (Category : Preface_specs.CATEGORY) -> functor (Core : Preface_specs.Arrow_choice.CORE with type ('a, 'b) t = ('a, 'b) Category.t) -> functor (Operation : Preface_specs.Arrow_choice.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t) -> Preface_specs.Arrow_choice.INFIX with type ('a, 'b) t = ('a, 'b) Core.t