Parameter Over_category_and_via_arrow_and_fst_and_left.2-Req

type ('a, 'b) t

The type held by the Arrow_choice.

val left : ('a, 'b) t -> (('a, 'c) Preface_core.Shims.Either.t, ('b, 'c) Preface_core.Shims.Either.t) t

Feed marked inputs through the argument arrow, passing the rest through unchanged to the output.

type ('a, 'b) t

The type held by the Arrow.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val fst : ('a, 'b) t -> ('a * 'd, 'b * 'd) t

Send the first component of the input through the argument Arrow, and copy the rest unchanged to the output.