Parameter Over_over_category_and_via_arrow_and_split_and_left.2-Req

type ('a, 'b) t = ('a, 'b) Category.t

The type held by the Arrow_choice.

val left : ('a, 'b) t -> (('a, 'c) Preface_core.Shims.Either.t, ('b, 'c) Preface_core.Shims.Either.t) t

Feed marked inputs through the argument arrow, passing the rest through unchanged to the output.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val split : ('a, 'b) t -> ('c, 'd) t -> ('a * 'c, 'b * 'd) t

Split the input between the two given Arrows and combine their output.