Parameter Alias.Operation

type ('a, 'b) t

The type held by the Arrow.

val compose_right_to_left : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

An alias of CORE.compose.

val compose_left_to_right : ('a, 'b) t -> ('b, 'c) t -> ('a, 'c) t

An alias of CORE.compose with flipped argument.

val return : unit -> ('a, 'a) t

Represent the identity Arrow.

val snd : ('a, 'b) t -> ('d * 'a, 'd * 'b) t

Send the second component of the input through the given Arrow, and copy the rest unchanged to the output.

val fan_out : ('a, 'b) t -> ('a, 'c) t -> ('a, 'b * 'c) t

Send the input to both argument arrows and combine their output.

val pre_compose_left_to_right : ('a -> 'b) -> ('b, 'c) t -> ('a, 'c) t

Pre composition with a function (the function should be pure).

val post_compose_left_to_right : ('a, 'b) t -> ('b -> 'c) -> ('a, 'c) t

Post composition with a function (the function should be pure).

val pre_compose_right_to_left : ('b, 'c) t -> ('a -> 'b) -> ('a, 'c) t

Reversed version of Preface_specs.Arrow.OPERATION.pre_compose_left_to_right.

val post_compose_right_to_left : ('b -> 'c) -> ('a, 'b) t -> ('a, 'c) t

Reversed version of Preface_specs.Arrow.OPERATION.post_compose_left_to_right.

val times_nel : int -> ('a, 'b) t -> ('a, 'b) t option

times n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : ('a, 'b) t Preface_core.Nonempty_list.t -> ('a, 'b) t

Reduce a Nonempty_list.t using combine.

val times : int -> ('a, 'b) t -> ('a, 'b) t

times_nel n x apply combine on x n times. If n is lower than 1 the function will returns neutral.

val reduce : ('a, 'b) t list -> ('a, 'b) t

Reduce a List.t using combine.