Module Preface_make.Arrow_zero

Building a Preface_specs.Arrow_zero

Using the minimal definition

Using neutral over an Arrow

Build an Preface_specs.ARROW_ZERO using Preface_specs.Arrow_zero.WITH_NEUTRAL over an Preface_specs.ARROW.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_arrow : functor (Arrow : Preface_specs.ARROW) -> functor (Req : Preface_specs.Arrow_zero.WITH_NEUTRAL with type ('a, 'b) t = ('a, 'b) Arrow.t) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = ('a, 'b) Req.t

Using neutral, arrow and fst over a Category

Build an Preface_specs.ARROW_ZERO using Preface_specs.Arrow_zero.WITH_ARROW_AND_FST over a Preface_specs.CATEGORY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_category_and_via_arrow_and_fst : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_zero.WITH_ARROW_AND_FST with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = ('a, 'b) Req.t

Using neutral, arrow and split over a Category

Build an Preface_specs.ARROW_ZERO using Preface_specs.Arrow_zero.WITH_ARROW_AND_SPLIT over a Preface_specs.CATEGORY.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_category_and_via_arrow_an_split : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_zero.WITH_ARROW_AND_SPLIT with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = ('a, 'b) Req.t

Arrow Zero Algebra

Construction of Preface_specs.ARROW_ZERO by combining them.

Product

Construct the product of two Preface_specs.ARROW_ZERO.

module Product : functor (F : Preface_specs.ARROW_ZERO) -> functor (G : Preface_specs.ARROW_ZERO) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = ('a, 'b) F.t * ('a, 'b) G.t

From other abstraction

From an Arrow Plus

Produces an Preface_specs.ARROW_ZERO from an Preface_specs.ARROW_PLUS.

module From_arrow_plus : functor (Plus : Preface_specs.ARROW_PLUS) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = ('a, 'b) Plus.t

From a Monad Plus

Produces an Preface_specs.ARROW_ZERO from a Preface_specs.MONAD_PLUS (using the Kleisli Arrow).

module From_monad_plus : functor (Monad : Preface_specs.Monad_plus.CORE) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = 'a -> 'b Monad.t

Manual construction

Advanced way to build an Preface_specs.ARROW_ZERO, constructing and assembling a component-by-component of Preface_specs.ARROW_ZERO. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Arrow_zero.CORE) -> functor (Operation : Preface_specs.Arrow_zero.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t) -> functor (Alias : Preface_specs.Arrow_zero.ALIAS with type ('a, 'b) t = ('a, 'b) Operation.t) -> functor (Infix : Preface_specs.Arrow_zero.INFIX with type ('a, 'b) t = ('a, 'b) Alias.t) -> Preface_specs.ARROW_ZERO with type ('a, 'b) t = ('a, 'b) Infix.t

Building Core

module Core_over_category_and_via_arrow_and_fst : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_zero.WITH_ARROW_AND_FST with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_zero.CORE with type ('a, 'b) t = ('a, 'b) Req.t

module Core_over_category_and_via_arrow_and_split : functor (Category : Preface_specs.CATEGORY) -> functor (Req : Preface_specs.Arrow_zero.WITH_ARROW_AND_SPLIT with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_zero.CORE with type ('a, 'b) t = ('a, 'b) Req.t

Deriving Operation

module Operation_over_category : functor (Category : Preface_specs.CATEGORY) -> functor (Core : Preface_specs.Arrow_zero.CORE with type ('a, 'b) t = ('a, 'b) Category.t) -> Preface_specs.Arrow_zero.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t

Deriving Alias

module Alias : functor (Operation : Preface_specs.Arrow_zero.OPERATION) -> Preface_specs.Arrow_zero.ALIAS with type ('a, 'b) t = ('a, 'b) Operation.t

Deriving Infix

module Infix_over_category : functor (Category : Preface_specs.CATEGORY) -> functor (Core : Preface_specs.Arrow_zero.CORE with type ('a, 'b) t = ('a, 'b) Category.t) -> functor (Operation : Preface_specs.Arrow_zero.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t) -> Preface_specs.Arrow_zero.INFIX with type ('a, 'b) t = ('a, 'b) Operation.t