Module Bounded_join_semilattice.Core_over_join_semilattice_and_via_bottom

Parameters

module Join_req : Preface_specs.Join_semilattice.CORE

module Req : Preface_specs.Bounded_join_semilattice.WITH_BOTTOM with type t = Join_req.t

Signature

include Preface_specs.Bounded_join_semilattice.WITH_BOTTOM with type t = Req.t

type t = Req.t

The type held by the Bounded_join_semilattice

val bottom : t

The bottom value represent the least element of the join semilattice

include Preface_specs.Join_semilattice.WITH_JOIN with type t := t

val join : t -> t -> t

join x y is the least upper bound of {x,y}.