Module Bounded_join_semilattice.Via

Parameters

module Core : Preface_specs.Bounded_join_semilattice.CORE

module Infix : Preface_specs.Bounded_join_semilattice.INFIX with type t = Core.t

Signature

type t = Infix.t

The type hold by the Bounded_join_semilattice

Functions

include Preface_specs.Bounded_join_semilattice.WITH_BOTTOM with type t := t

val bottom : t

The bottom value represent the least element of the join semilattice

include Preface_specs.Join_semilattice.WITH_JOIN with type t := t

val join : t -> t -> t

join x y is the least upper bound of {x,y}.

Infix operators

module Infix : Preface_specs.Bounded_join_semilattice.INFIX with type t = t

val (||) : t -> t -> t

Infix version of join.