Module Bounded_lattice.Core_via_bounded_join_lattice_and_bounded_meet_lattice

Parameters

module Req : Preface_specs.Bounded_lattice.WITH_BOUNDED_JOIN_LATTICE_AND_BOUNDED_MEET_LATTICE

Signature

include Preface_specs.Bounded_join_semilattice.WITH_JOIN_AND_BOTTOM with type t = Req.t

include Preface_specs.Bounded_join_semilattice.WITH_BOTTOM with type t = Req.t

type t = Req.t

The type held by the Bounded_join_semilattice

val bottom : t

The bottom value represent the least element of the join semilattice

include Preface_specs.Join_semilattice.WITH_JOIN with type t := t

val join : t -> t -> t

join x y is the least upper bound of {x,y}.

include Preface_specs.Bounded_meet_semilattice.WITH_MEET_AND_TOP with type t := t

include Preface_specs.Bounded_meet_semilattice.WITH_TOP with type t := t

val top : t

the top value represent the greatest element of the meet semilattice

include Preface_specs.Meet_semilattice.WITH_MEET with type t := t

val meet : t -> t -> t

meet x y is the greatest lower bound of {x, y}.