Parameter Over_bounded_join_lattice_and_bounded_meet_lattice.Join_req

include Preface_specs.Bounded_join_semilattice.WITH_BOTTOM

type t

The type held by the Bounded_join_semilattice

val bottom : t

The bottom value represent the least element of the join semilattice

include Preface_specs.Join_semilattice.WITH_JOIN with type t := t

val join : t -> t -> t

join x y is the least upper bound of {x,y}.