Module Bounded_meet_semilattice.Via_meet_and_top

Parameters

module Req : Preface_specs.Bounded_meet_semilattice.WITH_MEET_AND_TOP

Signature

type t = Req.t

The type hold by the Bounded_meet_semilattice

Functions

include Preface_specs.Bounded_meet_semilattice.WITH_TOP with type t := t

val top : t

the top value represent the greatest element of the meet semilattice

include Preface_specs.Meet_semilattice.WITH_MEET with type t := t

val meet : t -> t -> t

meet x y is the greatest lower bound of {x, y}.

Infix operators

module Infix : Preface_specs.Bounded_meet_semilattice.INFIX with type t := t

val (&&) : t -> t -> t

Infix version of meet.