Module Preface_make.Category

Building a Preface_specs.Category

Using the minimal definition

Build a Preface_specs.CATEGORY using Preface_specs.Category.WITH_ID_AND_COMPOSE.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_id_and_compose : functor (Req : Preface_specs.Category.WITH_ID_AND_COMPOSE) -> Preface_specs.CATEGORY with type ('a, 'b) t = ('a, 'b) Req.t

Category Algebra

Construction of Preface_specs.CATEGORY by combining them.

Product

Construct the product of two Preface_specs.CATEGORY.

module Product : functor (F : Preface_specs.CATEGORY) -> functor (G : Preface_specs.CATEGORY) -> Preface_specs.CATEGORY with type ('a, 'b) t = ('a, 'b) F.t * ('a, 'b) G.t

From other abstraction

From a Monad

Produces a Preface_specs.CATEGORY from a Preface_specs.MONAD.

module From_monad : functor (Monad : Preface_specs.Monad.CORE) -> Preface_specs.CATEGORY with type ('a, 'b) t = 'a -> 'b Monad.t

Manual construction

Advanced way to build a Preface_specs.CATEGORY, constructing and assembling a component-by-component of Preface_specs.CATEGORY. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Category.CORE) -> functor (Operation : Preface_specs.Category.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t) -> functor (Infix : Preface_specs.Category.INFIX with type ('a, 'b) t = ('a, 'b) Operation.t) -> Preface_specs.CATEGORY with type ('a, 'b) t = ('a, 'b) Infix.t

Building Core

module Core : functor (Req : Preface_specs.Category.WITH_ID_AND_COMPOSE) -> Preface_specs.Category.CORE with type ('a, 'b) t = ('a, 'b) Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Category.CORE) -> Preface_specs.Category.OPERATION with type ('a, 'b) t = ('a, 'b) Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Category.CORE) -> Preface_specs.Category.INFIX with type ('a, 'b) t = ('a, 'b) Core.t