Parameter Via.Operation

type 'a t = 'a Core.t

The type held by the Comonad.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over t from 'a to 'b.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val compose_right_to_left : ('b t -> 'c) -> ('a t -> 'b) -> 'a t -> 'c

Composing co-monadic functions using Co-Kleisli Arrow (from right to left).

val replace : 'a -> 'b t -> 'a t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : 'a t -> unit t

Create a new t, replacing all values in the given functor by unit.