Module Free_applicative.Over_applicative

Parameters

A : Preface_specs.Applicative.CORE

Signature

include Preface_specs.FREE_APPLICATIVE with type 'a f = 'a A.t

type 'a f

The type held by the Preface_specs.Functor.

type _ t =

| Pure : 'a -> 'a t

| Apply : ('a -> 'b) t * 'a f -> 'b t

The type held by the Free applicative.

val promote : 'a f -> 'a t

Promote a value from the Preface_specs.Functor into the Free applicative.

module To_applicative : functor (Applicative : Preface_specs.Applicative.CORE) -> Preface_specs.Free_applicative.TO_APPLICATIVE with type 'a t := 'a t and type 'a f := 'a f and type 'a applicative := 'a Applicative.t

The natural transformation from a Free applicative to an other Preface_specs.Applicative.

module To_monoid : functor (Monoid : Preface_specs.Monoid.CORE) -> Preface_specs.Free_applicative.TO_MONOID with type 'a t := 'a t and type 'a f := 'a f and type monoid := Monoid.t

The natural transformation from a Free applicative to a Preface_specs.Monoid.

Applicative API

A Free applicative is also an Preface_specs.Applicative.

type 'a t

The type held by the Applicative.

Functions

type 'a t

The type held by the Applicative.

val pure : 'a -> 'a t

Lift a value from 'a into a new 'a t.

val apply : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over 'a t to 'b t.

type 'a t

The type held by the Applicative.

val pure : 'a -> 'a t

Lift a value from 'a into a new 'a t.

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val product : 'a t -> 'b t -> ('a * 'b) t

Product functor mapping from 'a t and 'b t to ('a * 'b) t.

type 'a t

The type held by the Applicative.

val pure : 'a -> 'a t

Lift a value from 'a into a new 'a t.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Mapping over from 'a and 'b to 'c over 'a t and 'b t to 'c t.

type 'a t

The type held by the Applicative.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Mapping over from 'a and 'b and 'c to 'd over 'a t and 'b t and 'c t to 'd t.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

Infix operators

module Infix : Preface_specs.Applicative.INFIX with type 'a t := 'a t

type 'a t

The type held by the Applicative.

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over 'a t to 'b t.

val (<**>) : 'a t -> ('a -> 'b) t -> 'b t

Flipped Applicative functor of ('a -> 'b) t over 'a t to 'b t.

val (*>) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (<*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

type 'a t

The type held by the Functor.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Preface_specs.Functor.CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Preface_specs.Functor.CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Preface_specs.Functor.OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Preface_specs.Functor.OPERATION.replace.

Syntax

module Syntax : Preface_specs.Applicative.SYNTAX with type 'a t := 'a t

type 'a t

The type held by the Applicative.

val let+ : 'a t -> ('a -> 'b) -> 'b t

Flipped mapping over from 'a to 'b over 'a t to 'b t.

val and+ : 'a t -> 'b t -> ('a * 'b) t

Product functor mapping from 'a t and 'b t to ('a * 'b) t.

val run : 'a t -> 'a f