Module Preface_make.Freer_selective

Building a Preface_specs.Freer_selective

Using the minimal definition

Build a Preface_specs.FREER_SELECTIVE over an arbitrary type with one type parameter ( Preface_specs.Types.T1).

module Over (T : Preface_specs.Types.T1) : Preface_specs.FREER_SELECTIVE with type 'a f = 'a T.t

Over a Selective

Build a Preface_specs.FREER_SELECTIVE over an Preface_specs.SELECTIVE with a Natural transformation which extracts the underlying value and a run function which is the Natural transformation from the underlying applicative using the identity.

module Over_selective (F : Preface_specs.SELECTIVE) : sig ... end