Module Indexed_alternative.Via

Parameters

module Core : Preface_specs.Indexed_alternative.CORE

module Operation : Preface_specs.Indexed_alternative.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t

module Infix : Preface_specs.Indexed_alternative.INFIX with type ('a, 'index) t = ('a, 'index) Core.t

module Syntax : Preface_specs.Indexed_alternative.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t

Signature

Type

type ('a, 'index) t = ('a, 'index) Core.t

The type held by the Indexed Alternative.

Functions

val combine : ('a, 'index) t -> ('a, 'index) t -> ('a, 'index) t

Combine two values of t into one.

val neutral : ('a, 'index) t

The neutral element of the t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_apply.WITH_APPLY with type ('a, 'index) t := ('a, 'index) t

val apply : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

May apply a function wrapped into an t to a value also wrapped into an t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_functor.WITH_MAP with type ('a, 'index) t := ('a, 'index) t

val map : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

include Preface_specs.Indexed_apply.WITH_PRODUCT with type ('a, 'index) t := ('a, 'index) t

val product : ('a, 'index) t -> ('b, 'index) t -> ('a * 'b, 'index) t

Monoidal product between two t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := ('a, 'index) t

val pure : 'a -> ('a, 'index) t

Lift a value from 'a into a new t.

include Preface_specs.Indexed_apply.WITH_LIFT2 with type ('a, 'index) t := ('a, 'index) t

val lift2 : ('a -> 'b -> 'c) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val lift3 : ('a -> 'b -> 'c -> 'd) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t -> ('d, 'index) t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val times_nel : int -> ('a, 'index) t -> ('a, 'index) t option

times_nel n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : ('a, 'index) t Preface_core.Nonempty_list.t -> ('a, 'index) t

Reduce a Nonempty_list.t using combine.

val replace : 'a -> ('b, 'index) t -> ('a, 'index) t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : ('a, 'index) t -> (unit, 'index) t

Create a new t, replacing all values in the given functor by unit.

val times : int -> ('a, 'index) t -> ('a, 'index) t

times n x apply combine on x n times. If n is lower than 1 the function will returns neutral.

val reduce : ('a, 'index) t list -> ('a, 'index) t

Reduce a List.t using combine.

Infix operators

module Infix : Preface_specs.Indexed_alternative.INFIX with type ('a, 'index) t := ('a, 'index) t

val (<*>) : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (<**>) : ('a, 'index) t -> ('a -> 'b, 'index) t -> ('b, 'index) t

Flipped Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (*>) : (unit, 'index) t -> ('a, 'index) t -> ('a, 'index) t

Discard the value of the first argument.

val (<*) : ('a, 'index) t -> (unit, 'index) t -> ('a, 'index) t

Discard the value of the second argument.

Infix version of CORE.map.

Flipped and infix version of CORE.map.

Flipped and infix version of OPERATION.replace.

val (<|>) : ('a, 'index) t -> ('a, 'index) t -> ('a, 'index) t

Infix version of CORE.combine

val (<$>) : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Infix version of CORE.map.

val (<&>) : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

Flipped and infix version of CORE.map.

val (<$) : 'a -> ('b, 'index) t -> ('a, 'index) t

Infix version of OPERATION.replace.

val ($>) : ('a, 'index) t -> 'b -> ('b, 'index) t

Flipped and infix version of OPERATION.replace.

Syntax

module Syntax : Preface_specs.Indexed_alternative.SYNTAX with type ('a, 'index) t := ('a, 'index) t

include Preface_specs.Indexed_apply.SYNTAX with type ('a, 'index) t := ('a, 'index) t

val let+ : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

let operator for mapping.

val and+ : ('a, 'index) t -> ('b, 'index) t -> ('a * 'b, 'index) t

and operator for the monoidal product.