Module Preface_make.Indexed_apply

Building a Preface_specs.INDEXED_APPLY

Using the minimal definition

Using map and apply

Build a Preface_specs.APPLY using Preface_specs.APPLY.WITH_MAP_AND_APPLY.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_apply (Req : Preface_specs.Indexed_apply.WITH_MAP_AND_APPLY) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Req.t

module Over_functor_via_apply (Functor : Preface_specs.Indexed_functor.WITH_MAP) (Req : Preface_specs.Indexed_apply.WITH_APPLY with type ('a, 'index) t = ('a, 'index) Functor.t) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Req.t

Using map and product

Build a Preface_specs.INDEXED_APPLY using Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_product (Req : Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Req.t

Using product over functor

Build a Preface_specs.INDEXED_APPLY using Preface_specs.Indexed_functor.WITH_MAP and Preface_specs.Indexed_apply.WITH_PRODUCT.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_functor_via_product (Functor : Preface_specs.Indexed_functor.WITH_MAP) (Req : Preface_specs.Indexed_apply.WITH_PRODUCT with type ('a, 'index) t = ('a, 'index) Functor.t) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Req.t

Using map and lift2

Build a Preface_specs.INDEXED_APPLY using Preface_specs.Indexed_apply.WITH_LIFT2.

Other standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_lift2 (Req : Preface_specs.Indexed_apply.WITH_MAP_AND_LIFT2) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Req.t

Using lift2 over functor

Build a Preface_specs.INDEXED_APPLY using Preface_specs.Indexed_functor.WITH_MAP and Preface_specs.Indexed_apply.WITH_LIFT2.

Other standard method, using the minimal definition of an alt to derive its full API.

module Over_functor_via_lift2 (Functor : Preface_specs.Indexed_functor.WITH_MAP) (Req : Preface_specs.Indexed_apply.WITH_LIFT2 with type ('a, 'index) t = ('a, 'index) Functor.t) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Req.t

Manual construction

Advanced way to build an Preface_specs.INDEXED_APPLY, constructing and assembling a component-by-component of Preface_specs.INDEXED_APPLY. (In order to provide your own implementation for some features.)

Grouping of all components

module Via (Core : Preface_specs.Indexed_apply.CORE) (Operation : Preface_specs.Indexed_apply.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t) (Infix : Preface_specs.Indexed_apply.INFIX with type ('a, 'index) t = ('a, 'index) Core.t) (Syntax : Preface_specs.Indexed_apply.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t) : Preface_specs.INDEXED_APPLY with type ('a, 'index) t = ('a, 'index) Core.t

Building Core

module Core_via_map_and_apply (Req : Preface_specs.Indexed_apply.WITH_MAP_AND_APPLY) : Preface_specs.Indexed_apply.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_map_and_product (Req : Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT) : Preface_specs.Indexed_apply.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_map_and_lift2 (Req : Preface_specs.Indexed_apply.WITH_MAP_AND_LIFT2) : Preface_specs.Indexed_apply.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_over_functor_via_apply (Functor : Preface_specs.Indexed_functor.WITH_MAP) (Req : Preface_specs.Indexed_apply.WITH_APPLY with type ('a, 'index) t = ('a, 'index) Functor.t) : Preface_specs.Indexed_apply.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_over_functor_via_product (Functor : Preface_specs.Indexed_functor.WITH_MAP) (Req : Preface_specs.Indexed_apply.WITH_PRODUCT with type ('a, 'index) t = ('a, 'index) Functor.t) : Preface_specs.Indexed_apply.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_over_functor_via_lift2 (Functor : Preface_specs.Indexed_functor.WITH_MAP) (Req : Preface_specs.Indexed_apply.WITH_LIFT2 with type ('a, 'index) t = ('a, 'index) Functor.t) : Preface_specs.Indexed_apply.CORE with type ('a, 'index) t = ('a, 'index) Req.t

Deriving Operation

module Operation (Core : Preface_specs.Indexed_apply.CORE) : Preface_specs.Indexed_apply.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t

Deriving Syntax

module Syntax (Core : Preface_specs.Indexed_apply.CORE) : Preface_specs.Indexed_apply.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t

Deriving Infix

module Infix (Core : Preface_specs.Indexed_apply.CORE) (Operation : Preface_specs.Indexed_apply.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t) : Preface_specs.Indexed_apply.INFIX with type ('a, 'index) t = ('a, 'index) Core.t