Module Indexed_apply.Via_map_and_product

Parameters

module Req : Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT

Signature

Type

type ('a, 'index) t = ('a, 'index) Req.t

The type held by the Indexed Apply.

Functions

val map : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val apply : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

May apply a function wrapped into an t to a value also wrapped into an t.

val product : ('a, 'index) t -> ('b, 'index) t -> ('a * 'b, 'index) t

Monoidal product between two t.

val lift2 : ('a -> 'b -> 'c) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val lift3 : ('a -> 'b -> 'c -> 'd) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t -> ('d, 'index) t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val replace : 'a -> ('b, 'index) t -> ('a, 'index) t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : ('a, 'index) t -> (unit, 'index) t

Create a new t, replacing all values in the given functor by unit.

Infix operators

module Infix : Preface_specs.Indexed_apply.INFIX with type ('a, 'index) t := ('a, 'index) t

val (<*>) : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (<**>) : ('a, 'index) t -> ('a -> 'b, 'index) t -> ('b, 'index) t

Flipped Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (*>) : (unit, 'index) t -> ('a, 'index) t -> ('a, 'index) t

Discard the value of the first argument.

val (<*) : ('a, 'index) t -> (unit, 'index) t -> ('a, 'index) t

Discard the value of the second argument.

val (<$>) : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Infix version of CORE.map.

val (<&>) : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

Flipped and infix version of CORE.map.

val (<$) : 'a -> ('b, 'index) t -> ('a, 'index) t

Infix version of OPERATION.replace.

val ($>) : ('a, 'index) t -> 'b -> ('b, 'index) t

Flipped and infix version of OPERATION.replace.

Syntax

module Syntax : Preface_specs.Indexed_apply.SYNTAX with type ('a, 'index) t := ('a, 'index) t

val let+ : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

let operator for mapping.

val and+ : ('a, 'index) t -> ('b, 'index) t -> ('a * 'b, 'index) t

and operator for the monoidal product.