Module Preface_make.Indexed_foldable

Building an Preface_specs.INDEXED_FOLDABLE

Using the minimal definition

Using fold_map

Build a Preface_specs.INDEXED_FOLDABLE using Preface_specs.Indexed_foldable.WITH_FOLD_MAP.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_fold_map (Req : Preface_specs.Indexed_foldable.WITH_FOLD_MAP) : Preface_specs.INDEXED_FOLDABLE with type ('a, 'index) t = ('a, 'index) Req.t

Using fold_right

Build a Preface_specs.INDEXED_FOLDABLE using Preface_specs.Indexed_foldable.WITH_FOLD_RIGHT.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_fold_right (Req : Preface_specs.Indexed_foldable.WITH_FOLD_RIGHT) : Preface_specs.INDEXED_FOLDABLE with type ('a, 'index) t = ('a, 'index) Req.t

Manual construction

Advanced way to build a Preface_specs.INDEXED_FOLDABLE, constructing and assembling a component-by-component of Preface_specs.INDEXED_FOLDABLE. (In order to provide your own implementation for some features.)

Grouping of all components

module Via (C : Preface_specs.Indexed_foldable.CORE) (O : Preface_specs.Indexed_foldable.OPERATION with type ('a, 'index) t = ('a, 'index) C.t) : Preface_specs.INDEXED_FOLDABLE with type ('a, 'index) t = ('a, 'index) O.t

Building Core

module Core_via_fold_right (Req : Preface_specs.Indexed_foldable.WITH_FOLD_RIGHT) : Preface_specs.Indexed_foldable.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_fold_map (Req : Preface_specs.Indexed_foldable.WITH_FOLD_MAP) : Preface_specs.Indexed_foldable.CORE with type ('a, 'index) t = ('a, 'index) Req.t

Deriving Operation

module Operation (C : Preface_specs.Indexed_foldable.CORE) : Preface_specs.Indexed_foldable.OPERATION with type ('a, 'index) t = ('a, 'index) C.t