Module Indexed_functor.Via

Parameters

module Core : Preface_specs.Indexed_functor.CORE

module Operation : Preface_specs.Indexed_functor.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t

module Infix : Preface_specs.Indexed_functor.INFIX with type ('a, 'index) t = ('a, 'index) Core.t

module Syntax : Preface_specs.Indexed_functor.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t

Signature

Type

type ('a, 'index) t = ('a, 'index) Core.t

The type held by the Indexed Functor.

Functions

val map : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val replace : 'a -> ('b, 'index) t -> ('a, 'index) t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : ('a, 'index) t -> (unit, 'index) t

Create a new t, replacing all values in the given functor by unit.

Infix operators

module Infix : Preface_specs.Indexed_functor.INFIX with type ('a, 'index) t := ('a, 'index) t

val (<$>) : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Infix version of CORE.map.

val (<&>) : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

Flipped and infix version of CORE.map.

val (<$) : 'a -> ('b, 'index) t -> ('a, 'index) t

Infix version of OPERATION.replace.

val ($>) : ('a, 'index) t -> 'b -> ('b, 'index) t

Flipped and infix version of OPERATION.replace.

Syntax operators

module Syntax : Preface_specs.Indexed_functor.SYNTAX with type ('a, 'index) t := ('a, 'index) t

val let+ : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

let operator for mapping.