Module Preface_make.Indexed_monad

Building a Preface_specs.Indexed_monad

Using the minimal definition

Using return and bind

Build a Preface_specs.INDEXED_MONAD using Preface_specs.Indexed_monad.WITH_RETURN_AND_BIND.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_return_and_bind (Req : Preface_specs.Indexed_monad.WITH_RETURN_AND_BIND) : Preface_specs.INDEXED_MONAD with type ('a, 'index) t = ('a, 'index) Req.t

Using return, map and join

Build a Preface_specs.INDEXED_MONAD using Preface_specs.Indexed_monad.WITH_RETURN_MAP_AND_JOIN.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_return_map_and_join (Req : Preface_specs.Indexed_monad.WITH_RETURN_MAP_AND_JOIN) : Preface_specs.INDEXED_MONAD with type ('a, 'index) t = ('a, 'index) Req.t

Using return and the kleisli composition

Build a Preface_specs.INDEXED_MONAD using Preface_specs.Indexed_monad.WITH_RETURN_AND_KLEISLI_COMPOSITION.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_return_and_kleisli_composition (Req : Preface_specs.Indexed_monad.WITH_RETURN_AND_KLEISLI_COMPOSITION) : Preface_specs.INDEXED_MONAD with type ('a, 'index) t = ('a, 'index) Req.t

Grouping of all components

module Via (Core : Preface_specs.Indexed_monad.CORE) (Operation : Preface_specs.Indexed_monad.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t) (Infix : Preface_specs.Indexed_monad.INFIX with type ('a, 'index) t = ('a, 'index) Core.t) (Syntax : Preface_specs.Indexed_monad.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t) : Preface_specs.INDEXED_MONAD with type ('a, 'index) t = ('a, 'index) Core.t

Building Core

module Core_via_return_and_bind (Req : Preface_specs.Indexed_monad.WITH_RETURN_AND_BIND) : Preface_specs.Indexed_monad.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_return_map_and_join (Req : Preface_specs.Indexed_monad.WITH_RETURN_MAP_AND_JOIN) : Preface_specs.Indexed_monad.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_return_and_kleisli_composition (Req : Preface_specs.Indexed_monad.WITH_RETURN_AND_KLEISLI_COMPOSITION) : Preface_specs.Indexed_monad.CORE with type ('a, 'index) t = ('a, 'index) Req.t

Deriving Operation

module Operation (Core : Preface_specs.Indexed_monad.CORE) : Preface_specs.Indexed_monad.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t

Deriving Syntax

module Syntax (Core : Preface_specs.Indexed_monad.CORE) : Preface_specs.Indexed_monad.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t

Deriving Infix

module Infix (Core : Preface_specs.Indexed_monad.CORE) (Operation : Preface_specs.Indexed_monad.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t) : Preface_specs.Indexed_monad.INFIX with type ('a, 'index) t = ('a, 'index) Core.t