Module Preface_make.Indexed_monad_plus

Building a Preface_specs.Indexed_monad_plus

Using the minimal definition

Using return, bind, neutral and combine

Build a Preface_specs.INDEXED_MONAD_PLUS using Preface_specs.Indexed_monad_plus.WITH_BIND.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_bind (Req : Preface_specs.Indexed_monad_plus.WITH_BIND) : Preface_specs.INDEXED_MONAD_PLUS with type ('a, 'index) t = ('a, 'index) Req.t

Using return, map, join, neutral and combine

Build a Preface_specs.INDEXED_MONAD_PLUS using Preface_specs.Indexed_onad_plus.WITH_MAP_AND_JOIN.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_join (Req : Preface_specs.Indexed_monad_plus.WITH_MAP_AND_JOIN) : Preface_specs.INDEXED_MONAD_PLUS with type ('a, 'index) t = ('a, 'index) Req.t

Using return, neutral, combine and the kleisli composition

Build a Preface_specs.INDEXED_MONAD_PLUS using Preface_specs.Indexed_monad_plus.WITH_KLEISLI_COMPOSITION.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_kleisli_composition (Req : Preface_specs.Indexed_monad_plus.WITH_KLEISLI_COMPOSITION) : Preface_specs.INDEXED_MONAD_PLUS with type ('a, 'index) t = ('a, 'index) Req.t

Over a Monad and an Alternative

Build a Preface_specs.INDEXED_MONAD_PLUS over a Preface_specs.INDEXED_MONAD and an neutral and combine from an Preface_specs.INDEXED_ALTERNATIVE.

module Over_monad_and_alternative (Monad : Preface_specs.INDEXED_MONAD) (Alternative : Preface_specs.INDEXED_ALTERNATIVE with type ('a, 'index) t = ('a, 'index) Monad.t) : Preface_specs.INDEXED_MONAD_PLUS with type ('a, 'index) t = ('a, 'index) Alternative.t

Over a Monad using neutral and combine

Build a Preface_specs.INDEXED_MONAD_PLUS over a Preface_specs.INDEXED_MONAD and using Preface_specs.Indexed_monad_plus.WITH_NEUTRAL_AND_COMBINE.

module Over_monad (Monad : Preface_specs.INDEXED_MONAD) (Req : Preface_specs.Indexed_monad_plus.WITH_NEUTRAL_AND_COMBINE with type ('a, 'index) t = ('a, 'index) Monad.t) : Preface_specs.INDEXED_MONAD_PLUS with type ('a, 'index) t = ('a, 'index) Req.t

Manual construction

Advanced way to build a Preface_specs.INDEXED_MONAD_PLUS, constructing and assembling a component-by-component of Preface_specs.INDEXED_MONAD_PLUS. (In order to provide your own implementation for some features.)

Grouping of all components

module Via (Core : Preface_specs.Indexed_monad_plus.CORE) (Operation : Preface_specs.Indexed_monad_plus.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t) (Infix : Preface_specs.Indexed_monad_plus.INFIX with type ('a, 'index) t = ('a, 'index) Core.t) (Syntax : Preface_specs.Indexed_monad_plus.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t) : Preface_specs.INDEXED_MONAD_PLUS with type ('a, 'index) t = ('a, 'index) Core.t

Building Core

module Core_via_bind (Req : Preface_specs.Indexed_monad_plus.WITH_BIND) : Preface_specs.Indexed_monad_plus.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_map_and_join (Req : Preface_specs.Indexed_monad_plus.WITH_MAP_AND_JOIN) : Preface_specs.Indexed_monad_plus.CORE with type ('a, 'index) t = ('a, 'index) Req.t

module Core_via_kleisli_composition (Req : Preface_specs.Indexed_monad_plus.WITH_KLEISLI_COMPOSITION) : Preface_specs.Indexed_monad_plus.CORE with type ('a, 'index) t = ('a, 'index) Req.t

Deriving Operation

module Operation (Core : Preface_specs.Indexed_monad_plus.CORE) : Preface_specs.Indexed_monad_plus.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t

Deriving Syntax

module Syntax (Core : Preface_specs.Indexed_monad_plus.CORE) : Preface_specs.Indexed_monad_plus.SYNTAX with type ('a, 'index) t = ('a, 'index) Core.t

Deriving Infix

module Infix (Core : Preface_specs.Indexed_monad_plus.CORE) (Operation : Preface_specs.Indexed_monad_plus.OPERATION with type ('a, 'index) t = ('a, 'index) Core.t) : Preface_specs.Indexed_monad_plus.INFIX with type ('a, 'index) t = ('a, 'index) Core.t