Parameter Infix.Operation

type ('a, 'index) t = ('a, 'index) Core.t

The type held by the Indexed Monad Plus.

val compose_right_to_left : ('b -> ('c, 'index) t) -> ('a -> ('b, 'index) t) -> 'a -> ('c, 'index) t

Composing monadic functions using Kleisli Arrow (from right to left).

val lift : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val lift2 : ('a -> 'b -> 'c) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift3 : ('a -> 'b -> 'c -> 'd) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t -> ('d, 'index) t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val times_nel : int -> ('a, 'index) t -> ('a, 'index) t option

times_nel n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : ('a, 'index) t Preface_core.Nonempty_list.t -> ('a, 'index) t

Reduce a Nonempty_list.t using combine.

val replace : 'a -> ('b, 'index) t -> ('a, 'index) t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : ('a, 'index) t -> (unit, 'index) t

Create a new t, replacing all values in the given functor by unit.

val times : int -> ('a, 'index) t -> ('a, 'index) t

times n x apply combine on x n times. If n is lower than 1 the function will returns neutral.

val reduce : ('a, 'index) t list -> ('a, 'index) t

Reduce a List.t using combine.

val filter : ('a -> bool) -> ('a, 'index) t -> ('a, 'index) t

Filtering over Monad_plus.