Parameter Core_via_join_and_meet.Req

include Preface_specs.Join_semilattice.WITH_JOIN

type t

The type held by the Join_semilattice.

val join : t -> t -> t

join x y is the least upper bound of {x,y}.

include Preface_specs.Meet_semilattice.WITH_MEET with type t := t

val meet : t -> t -> t

meet x y is the greatest lower bound of {x, y}.