Module Preface_make.Monad

Building a Preface_specs.Monad

Using the minimal definition

Using return and bind

Build a Preface_specs.MONAD using Preface_specs.Monad.WITH_BIND.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_bind (Req : Preface_specs.Monad.WITH_BIND) : Preface_specs.MONAD with type 'a t = 'a Req.t

Using return, map and join

Build a Preface_specs.MONAD using Preface_specs.Monad.WITH_MAP_AND_JOIN.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_join (Req : Preface_specs.Monad.WITH_MAP_AND_JOIN) : Preface_specs.MONAD with type 'a t = 'a Req.t

Using return and the kleisli composition

Build a Preface_specs.MONAD using Preface_specs.Monad.WITH_KLEISLI_COMPOSITION.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_kleisli_composition (Req : Preface_specs.Monad.WITH_KLEISLI_COMPOSITION) : Preface_specs.MONAD with type 'a t = 'a Req.t

Monad Algebra

Construction of Preface_specs.MONAD by combining them.

Product

Construct the product of two Preface_specs.MONAD.

module Product (F : Preface_specs.MONAD) (G : Preface_specs.MONAD) : Preface_specs.MONAD with type 'a t = 'a F.t * 'a G.t

From other abstraction

From a Monad Plus

Produces a Preface_specs.MONAD from a Preface_specs.MONAD_PLUS.

module From_monad_plus (Monad_plus : Preface_specs.MONAD_PLUS) : Preface_specs.MONAD with type 'a t = 'a Monad_plus.t

From an Arrow Apply

Produces a Preface_specs.MONAD from an Preface_specs.ARROW_APPLY.

module From_arrow_apply (A : Preface_specs.ARROW_APPLY) : Preface_specs.MONAD with type 'a t = (unit, 'a) A.t

Manual construction

Advanced way to build a Preface_specs.MONAD, constructing and assembling a component-by-component of Preface_specs.MONAD. (In order to provide your own implementation for some features.)

Grouping of all components

module Via (Core : Preface_specs.Monad.CORE) (Operation : Preface_specs.Monad.OPERATION with type 'a t = 'a Core.t) (Infix : Preface_specs.Monad.INFIX with type 'a t = 'a Core.t) (Syntax : Preface_specs.Monad.SYNTAX with type 'a t = 'a Core.t) : Preface_specs.MONAD with type 'a t = 'a Core.t

Building Core

module Core_via_bind (Req : Preface_specs.Monad.WITH_BIND) : Preface_specs.Monad.CORE with type 'a t = 'a Req.t

module Core_via_map_and_join (Req : Preface_specs.Monad.WITH_MAP_AND_JOIN) : Preface_specs.Monad.CORE with type 'a t = 'a Req.t

module Core_via_kleisli_composition (Req : Preface_specs.Monad.WITH_KLEISLI_COMPOSITION) : Preface_specs.Monad.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation (Core : Preface_specs.Monad.CORE) : Preface_specs.Monad.OPERATION with type 'a t = 'a Core.t

Deriving Syntax

module Syntax (Core : Preface_specs.Monad.CORE) : Preface_specs.Monad.SYNTAX with type 'a t = 'a Core.t

Deriving Infix

module Infix (Core : Preface_specs.Monad.CORE) (Operation : Preface_specs.Monad.OPERATION with type 'a t = 'a Core.t) : Preface_specs.Monad.INFIX with type 'a t = 'a Core.t