Parameter Operation.Core

type 'a t

The type held by the Monad.

include Preface_specs.Indexed_monad.WITH_RETURN with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_bind.WITH_BIND with type ('a, 'index) t := 'a t

val bind : ('a -> 'b t) -> 'a t -> 'b t

bind f m passes the result of computation m to function f.

include Preface_specs.Indexed_monad.WITH_RETURN with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_bind.WITH_MAP_AND_JOIN with type ('a, 'index) t := 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b.

val join : 'a t t -> 'a t

join remove one level of monadic structure, projecting its bound argument into the outer level.

include Preface_specs.Indexed_monad.WITH_RETURN with type ('a, 'index) t := 'a t

val return : 'a -> 'a t

Lift a value into a t.

include Preface_specs.Indexed_bind.WITH_KLEISLI_COMPOSITION with type ('a, 'index) t := 'a t

val compose_left_to_right : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from left to right).