Module Preface_make.Monoid

Building a Preface_specs.Monoid

Using the minimal definition

Using neutral and combine

Build a Preface_specs.MONOID using Preface_specs.Monoid.WITH_NEUTRAL_AND_COMBINE.

Standard method, using the minimal definition of a monoid to derive its full API.

module Via_combine_and_neutral : functor (Req : Preface_specs.Monoid.WITH_NEUTRAL_AND_COMBINE) -> Preface_specs.MONOID with type t = Req.t

Over a Semigroup

Build a Preface_specs.MONOID over a Preface_specs.SEMIGROUP.

If you already have a Semigroup, you can enrich it by passing only the neutral function.

module Over_semigroup : functor (S : Preface_specs.SEMIGROUP) -> functor (M : Preface_specs.Monoid.WITH_NEUTRAL with type t = S.t) -> Preface_specs.MONOID with type t = S.t

From other abstraction

From an Alternative

Specialize an Preface_specs.ALTERNATIVE into a Preface_specs.MONOID.

module From_alternative : functor (Alternative : Preface_specs.ALTERNATIVE) -> functor (T : Preface_specs.Types.T0) -> Preface_specs.MONOID with type t = T.t Alternative.t

From a Monad plus

Specialize an Preface_specs.MONAD_PLUS into a Preface_specs.MONOID.

module From_monad_plus : functor (Monad_plus : Preface_specs.MONAD_PLUS) -> functor (T : Preface_specs.Types.T0) -> Preface_specs.MONOID with type t = T.t Monad_plus.t

Manual construction

Advanced way to build a Preface_specs.MONOID, constructing and assembling a component-by-component of Preface_specs.MONOID. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Monoid.CORE) -> functor (Operation : Preface_specs.Monoid.OPERATION with type t = Core.t) -> functor (Infix : Preface_specs.Monoid.INFIX with type t = Operation.t) -> Preface_specs.MONOID with type t = Infix.t

Building Core

module Core_over_semigroup : functor (S : Preface_specs.SEMIGROUP) -> functor (M : Preface_specs.Monoid.WITH_NEUTRAL with type t = S.t) -> Preface_specs.Monoid.CORE with type t = M.t

module Core : functor (Req : Preface_specs.Monoid.WITH_NEUTRAL_AND_COMBINE) -> Preface_specs.Monoid.CORE with type t = Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Monoid.CORE) -> Preface_specs.Monoid.OPERATION with type t = Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Monoid.CORE) -> Preface_specs.Monoid.INFIX with type t = Core.t