Module Preface_make.Selective

Building a Preface_specs.Selective

Using the minimal definition

Over an Applicative using select

Build a Preface_specs.SELECTIVE using Preface_specs.Selective.WITH_SELECT on top of an Preface_specs.APPLICATIVE.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_applicative_via_select (Applicative : Preface_specs.APPLICATIVE) (Req : Preface_specs.Selective.WITH_SELECT with type 'a t = 'a Applicative.t) : Preface_specs.SELECTIVE with type 'a t = 'a Req.t

Over an Applicative using branch

Build a Preface_specs.SELECTIVE using Preface_specs.Selective.WITH_BRANCH on top of an Preface_specs.APPLICATIVE.

Standard method, using the minimal definition of an alt to derive its full API.

module Over_applicative_via_branch (Applicative : Preface_specs.APPLICATIVE) (Req : Preface_specs.Selective.WITH_BRANCH with type 'a t = 'a Applicative.t) : Preface_specs.SELECTIVE with type 'a t = 'a Req.t

Over a Functor using select

Build a Preface_specs.SELECTIVE using Preface_specs.Selective.WITH_SELECT on top of an Preface_specs.FUNCTOR.

Standard method, using the minimal definition of an alt to derive its full API (including the Applicative API).

module Over_functor_via_select (Functor : Preface_specs.Functor.CORE) (Req : Preface_specs.Selective.WITH_PURE_AND_SELECT with type 'a t = 'a Functor.t) : Preface_specs.SELECTIVE with type 'a t = 'a Req.t

Over a Functor using branch

Build a Preface_specs.SELECTIVE using Preface_specs.Selective.WITH_BRANCH on top of an Preface_specs.FUNCTOR.

Standard method, using the minimal definition of an alt to derive its full API (including the Applicative API).

module Over_functor_via_branch (Functor : Preface_specs.Functor.CORE) (Req : Preface_specs.Selective.WITH_PURE_AND_BRANCH with type 'a t = 'a Functor.t) : Preface_specs.SELECTIVE with type 'a t = 'a Req.t

Selective Algebra

Construction of Preface_specs.SELECTIVE by combining them.

Composition

Right-to-left composition of Preface_specs.APPLICATIVE with Preface_specs.SELECTIVE.

module Composition (F : Preface_specs.APPLICATIVE) (G : Preface_specs.SELECTIVE) : Preface_specs.SELECTIVE with type 'a t = 'a G.t F.t

Product

Construct the product of two Preface_specs.SELECTIVE.

module Product (F : Preface_specs.SELECTIVE) (G : Preface_specs.SELECTIVE) : Preface_specs.SELECTIVE with type 'a t = 'a F.t * 'a G.t

From other abstraction

From an Arrow choice

Produces a Preface_specs.SELECTIVE from an Preface_specs.ARROW_CHOICE.

module From_arrow_choice (A : Preface_specs.ARROW_CHOICE) : Preface_specs.SELECTIVE with type 'a t = (unit, 'a) A.t

From a Monoid

Produces a Preface_specs.SELECTIVE from a Preface_specs.MONOID. This Selective is called Const or a phantom monoid.

module Const (M : Preface_specs.Monoid.CORE) : sig ... end

Manual construction

Advanced way to build a Preface_specs.SELECTIVE, constructing and assembling a component-by-component of Preface_specs.SELECTIVE. (In order to provide your own implementation for some features.)

Grouping of all components

module Via (Core : Preface_specs.Selective.CORE) (Operation : Preface_specs.Selective.OPERATION with type 'a t = 'a Core.t) (Infix : Preface_specs.Selective.INFIX with type 'a t = 'a Core.t) (Syntax : Preface_specs.Selective.SYNTAX with type 'a t = 'a Core.t) : Preface_specs.SELECTIVE with type 'a t = 'a Core.t

Building Core

module Core_over_functor_via_select (Functor : Preface_specs.Functor.CORE) (Req : Preface_specs.Selective.WITH_PURE_AND_SELECT with type 'a t = 'a Functor.t) : Preface_specs.Selective.CORE with type 'a t = 'a Req.t

module Core_over_functor_via_branch (Functor : Preface_specs.Functor.CORE) (Req : Preface_specs.Selective.WITH_PURE_AND_BRANCH with type 'a t = 'a Functor.t) : Preface_specs.Selective.CORE with type 'a t = 'a Req.t

module Core_over_applicative_via_select (Applicative : Preface_specs.Applicative.CORE) (Req : Preface_specs.Selective.WITH_SELECT with type 'a t = 'a Applicative.t) : Preface_specs.Selective.CORE with type 'a t = 'a Req.t

module Core_over_applicative_via_branch (Applicative : Preface_specs.Applicative.CORE) (Req : Preface_specs.Selective.WITH_BRANCH with type 'a t = 'a Applicative.t) : Preface_specs.Selective.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation (Core : Preface_specs.Selective.CORE) : Preface_specs.Selective.OPERATION with type 'a t = 'a Core.t

Deriving Infix

module Infix (Core : Preface_specs.Selective.CORE) (Operation : Preface_specs.Selective.OPERATION with type 'a t = 'a Core.t) : Preface_specs.Selective.INFIX with type 'a t = 'a Core.t

Deriving Syntax

module Syntax (Core : Preface_specs.Selective.CORE) : Preface_specs.Selective.SYNTAX with type 'a t = 'a Core.t

Deriving Select from a Monad

module Select_from_monad (Monad : Preface_specs.MONAD) : Preface_specs.Selective.WITH_SELECT with type 'a t = 'a Monad.t