Module Preface_make.Semigroup

Building a Preface_specs.Semigroup

Using the minimal definition

Build a Preface_specs.SEMIGROUP using Preface_specs.Semigroup.WITH_COMBINE.

Standard method, using the minimal definition of a semigroup to derive its full API.

module Via_combine : functor (Req : Preface_specs.Semigroup.WITH_COMBINE) -> Preface_specs.SEMIGROUP with type t = Req.t

From other abstraction

From an Alt

Specialize an Preface_specs.ALT into a Preface_specs.SEMIGROUP.

module From_alt : functor (Alt : Preface_specs.ALT) -> functor (T : Preface_specs.Types.T0) -> Preface_specs.SEMIGROUP with type t = T.t Alt.t

From an Alternative

Specialize an Preface_specs.ALTERNATIVE into a Preface_specs.SEMIGROUP.

module From_alternative : functor (Alternative : Preface_specs.ALTERNATIVE) -> functor (T : Preface_specs.Types.T0) -> Preface_specs.SEMIGROUP with type t = T.t Alternative.t

From a Monad plus

Specialize an Preface_specs.MONAD_PLUS into a Preface_specs.SEMIGROUP.

module From_monad_plus : functor (Monad_plus : Preface_specs.MONAD_PLUS) -> functor (T : Preface_specs.Types.T0) -> Preface_specs.SEMIGROUP with type t = T.t Monad_plus.t

Manual construction

Advanced way to build a Preface_specs.SEMIGROUP, constructing and assembling a component-by-component of Preface_specs.SEMIGROUP. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Semigroup.CORE) -> functor (Operation : Preface_specs.Semigroup.OPERATION with type t = Core.t) -> functor (Infix : Preface_specs.Semigroup.INFIX with type t = Operation.t) -> Preface_specs.SEMIGROUP with type t = Infix.t

Building Core

module Core : functor (Req : Preface_specs.Semigroup.WITH_COMBINE) -> Preface_specs.Semigroup.CORE with type t = Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Semigroup.CORE) -> Preface_specs.Semigroup.OPERATION with type t = Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Semigroup.CORE) -> Preface_specs.Semigroup.INFIX with type t = Core.t