Module Traversable.Core_over_monad

Parameters

module M : Preface_specs.MONAD

module Req : Preface_specs.Traversable.WITH_TRAVERSE with type 'a t = 'a M.t

Signature

type 'a t = 'a Req.t

The type held by the Traversable.

type 'a iter = 'a Req.iter

The iterable type held by the Traversable.

val traverse : ('a -> 'b t) -> 'a iter -> 'b iter t

Map each element of a structure to an action, evaluate these actions from left to right, and collect the results. *