Parameter Over_monad.2-Req

type 'a t = 'a M.t

The type held by the Traversable.

type 'a iter

The iterable type held by the Traversable.

val traverse : ('a -> 'b t) -> 'a iter -> 'b iter t

Map each element of a structure to an action, evaluate these actions from left to right, and collect the results. *