Module Preface_make__.Alt

Using the minimal definition

Using neutral and combine

Build a Preface_specs.ALT using Preface_specs.Alt.WITH_COMBINE_AND_MAP.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_combine : functor (Req : Preface_specs.Alt.WITH_COMBINE_AND_MAP) -> Preface_specs.ALT with type 'a t = 'a Req.t

Over a Functor

Build a Preface_specs.ALT over a Preface_specs.FUNCTOR.

If you already have a Functor, you can enrich it by passing only the combine function.

module Over_functor : functor (Functor : Preface_specs.FUNCTOR) -> functor (Combine : Preface_specs.Alt.WITH_COMBINE with type 'a t = 'a Functor.t) -> Preface_specs.ALT with type 'a t = 'a Combine.t

Functor Algebra

Construction of Preface_specs.ALT by combining them.

Composition

Right-to-left composition of Alt with Preface_specs.FUNCTOR.

module Composition : functor (F : Preface_specs.ALT) -> functor (G : Preface_specs.FUNCTOR) -> Preface_specs.ALT with type 'a t = 'a G.t F.t

Product

Construct the product of two Preface_specs.ALT.

module Product : functor (F : Preface_specs.ALT) -> functor (G : Preface_specs.ALT) -> Preface_specs.ALT with type 'a t = 'a F.t * 'a G.t

Manual construction

Advanced way to build an Preface_specs.ALT, constructing and assembling a component-by-component of Preface_specs.ALT. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Alt.CORE) -> functor (Operation : Preface_specs.Alt.OPERATION with type 'a t = 'a Core.t) -> functor (Infix : Preface_specs.Alt.INFIX with type 'a t = 'a Operation.t) -> Preface_specs.ALT with type 'a t = 'a Infix.t

Building Core

module Core_over_functor : functor (Functor : Preface_specs.FUNCTOR) -> functor (Req : Preface_specs.Alt.WITH_COMBINE with type 'a t = 'a Functor.t) -> Preface_specs.Alt.CORE with type 'a t = 'a Req.t

module Core : functor (Req : Preface_specs.Alt.WITH_COMBINE_AND_MAP) -> Preface_specs.Alt.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Alt.CORE) -> Preface_specs.Alt.OPERATION with type 'a t = 'a Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Alt.CORE) -> functor (Operation : Preface_specs.Alt.OPERATION with type 'a t = 'a Core.t) -> Preface_specs.Alt.INFIX with type 'a t = 'a Core.t