Parameter Via.2-Operation

type 'a t

The type held by the Applicative.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Mapping over from 'a and 'b and 'c to 'd over 'a t and 'b t and 'c t to 'd t.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

type 'a t

A type 'a t held by the Alt.

val times : int -> 'a t -> 'a t option

times n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : 'a t Preface_core.Nonempty_list.t -> 'a t

Reduce a Nonempty_list.t using combine.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

val reduce : 'a t list -> 'a t

Reduce a List.t using combine.