Module Preface_make__.Contravariant

Using the minimal definition

Build a Preface_specs.CONTRAVARIANT using Preface_specs.Contravariant.WITH_CONTRAMAP.

Standard method, using the minimal definition of a contravariant functor to derive its full API.

module Via_contramap : functor (Req : Preface_specs.Contravariant.WITH_CONTRAMAP) -> Preface_specs.CONTRAVARIANT with type 'a t = 'a Req.t

Contravariant Algebra

Construction of Preface_specs.CONTRAVARIANT by combining them.

Contravariant composition

Construction of Preface_specs.CONTRAVARIANT by left-to-right composition with Preface_specs.FUNCTOR.

module Composition : functor (F : Preface_specs.FUNCTOR) -> functor (G : Preface_specs.CONTRAVARIANT) -> Preface_specs.CONTRAVARIANT with type 'a t = 'a G.t F.t

Manual construction

Advanced way to build a Preface_specs.CONTRAVARIANT, constructing and assembling a component-by-component of Preface_specs.CONTRAVARIANT. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Contravariant.CORE) -> functor (Operation : Preface_specs.Contravariant.OPERATION with type 'a t = 'a Core.t) -> functor (Infix : Preface_specs.Contravariant.INFIX with type 'a t = 'a Operation.t) -> Preface_specs.CONTRAVARIANT with type 'a t = 'a Infix.t

Building Core

module Core : functor (Req : Preface_specs.Contravariant.WITH_CONTRAMAP) -> Preface_specs.Contravariant.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Contravariant.CORE) -> Preface_specs.Contravariant.OPERATION with type 'a t = 'a Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Contravariant.CORE) -> functor (Operation : Preface_specs.Contravariant.OPERATION with type 'a t = 'a Core.t) -> Preface_specs.Contravariant.INFIX with type 'a t = 'a Operation.t