Module Preface_make__.Functor

Using the minimal definition

Build a Preface_specs.FUNCTOR using Preface_specs.Functor.WITH_MAP.

Standard method, using the minimal definition of a functor to derive its full API.

module Via_map : functor (Req : Preface_specs.Functor.WITH_MAP) -> Preface_specs.FUNCTOR with type 'a t = 'a Req.t

Functor Algebra

Construction of Preface_specs.FUNCTOR by combining two other Preface_specs.FUNCTOR.

Composition

Right-to-left composition of Preface_specs.FUNCTOR.

module Composition : functor (F : Preface_specs.FUNCTOR) -> functor (G : Preface_specs.FUNCTOR) -> Preface_specs.FUNCTOR with type 'a t = 'a G.t F.t

Product

Construct the product of two Preface_specs.FUNCTOR.

module Product : functor (F : Preface_specs.FUNCTOR) -> functor (G : Preface_specs.FUNCTOR) -> Preface_specs.FUNCTOR with type 'a t = 'a F.t * 'a G.t

Sum

Sum of Preface_specs.FUNCTOR using the technique described in Data types à la carte by W. Swierstra.

module Sum : functor (F : Preface_specs.FUNCTOR) -> functor (G : Preface_specs.FUNCTOR) -> sig ... end

From other abstraction

From an Arrow

Specialize an Preface_specs.ARROW into a Preface_specs.FUNCTOR.

module From_arrow : functor (A : Preface_specs.ARROW) -> Preface_specs.FUNCTOR with type 'a t = (unit, 'a) A.t

From an Applicative

Specialize an Preface_specs.APPLICATIVE into a Preface_specs.FUNCTOR.

module From_applicative : functor (Applicative : Preface_specs.APPLICATIVE) -> Preface_specs.FUNCTOR with type 'a t = 'a Applicative.t

From an Alt

Specialize an Preface_specs.ALT into a Preface_specs.FUNCTOR.

module From_alt : functor (Alt : Preface_specs.ALT) -> Preface_specs.FUNCTOR with type 'a t = 'a Alt.t

From a Monad

Specialize a Preface_specs.MONAD into a Preface_specs.FUNCTOR.

module From_monad : functor (Monad : Preface_specs.MONAD) -> Preface_specs.FUNCTOR with type 'a t = 'a Monad.t

From an Alternative

Specialize an Preface_specs.ALTERNATIVE into a Preface_specs.FUNCTOR.

module From_alternative : functor (Alternative : Preface_specs.ALTERNATIVE) -> Preface_specs.FUNCTOR with type 'a t = 'a Alternative.t

From a Monad plus

Specialize a Preface_specs.MONAD_PLUS into a Preface_specs.FUNCTOR.

module From_monad_plus : functor (Monad_plus : Preface_specs.MONAD_PLUS) -> Preface_specs.FUNCTOR with type 'a t = 'a Monad_plus.t

From a Comonad

Specialize a Preface_specs.COMONAD into a Preface_specs.FUNCTOR.

module From_comonad : functor (Comonad : Preface_specs.COMONAD) -> Preface_specs.FUNCTOR with type 'a t = 'a Comonad.t

Manual construction

Advanced way to build a Preface_specs.FUNCTOR, constructing and assembling a component-by-component of Preface_specs.FUNCTOR. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Functor.CORE) -> functor (Operation : Preface_specs.Functor.OPERATION with type 'a t = 'a Core.t) -> functor (Infix : Preface_specs.Functor.INFIX with type 'a t = 'a Core.t) -> Preface_specs.FUNCTOR with type 'a t = 'a Core.t

Building Core

module Core : functor (Req : Preface_specs.Functor.WITH_MAP) -> Preface_specs.Functor.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Functor.CORE) -> Preface_specs.Functor.OPERATION with type 'a t = 'a Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Functor.CORE) -> functor (Operation : Preface_specs.Functor.OPERATION with type 'a t = 'a Core.t) -> Preface_specs.Functor.INFIX with type 'a t = 'a Core.t