Module Preface_make__Arrow.From_arrow_plus

Parameters

Plus : Preface_specs.ARROW_PLUS

Signature

Type

type ('a, 'b) t = ('a, 'b) Plus.t

The type held by the Arrow.

Functions

type ('a, 'b) t

The type held by the Arrow.

type ('a, 'b) t

The type held by the Category.

val id : ('a, 'a) t

The identity morphism.

val compose : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

Morphism composition (from right to left).

type ('a, 'b) t

The type held by the Arrow.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val fst : ('a, 'b) t -> ('a * 'd, 'b * 'd) t

Send the first component of the input through the argument Arrow, and copy the rest unchanged to the output.

type ('a, 'b) t

The type held by the Arrow.

val arrow : ('a -> 'b) -> ('a, 'b) t

Lift a function to an Arrow.

val split : ('a, 'b) t -> ('c, 'd) t -> ('a * 'c, 'b * 'd) t

Split the input between the two given Arrows and combine their output.

type ('a, 'b) t

The type held by the Arrow.

type ('a, 'b) t

The type held by the Category.

val compose_right_to_left : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

An alias of Category.CORE.compose.

val compose_left_to_right : ('a, 'b) t -> ('b, 'c) t -> ('a, 'c) t

An alias of Category.CORE.compose with flipped argument.

val return : unit -> ('a, 'a) t

Represent the identity Arrow.

val snd : ('a, 'b) t -> ('d * 'a, 'd * 'b) t

Send the second component of the input through the given Arrow, and copy the rest unchanged to the output.

val fan_out : ('a, 'b) t -> ('a, 'c) t -> ('a, 'b * 'c) t

Send the input to both argument arrows and combine their output.

val pre_compose_left_to_right : ('a -> 'b) -> ('b, 'c) t -> ('a, 'c) t

Precomposition with a function (the function should be pure).

val post_compose_left_to_right : ('a, 'b) t -> ('b -> 'c) -> ('a, 'c) t

Postcomposition with a function (the function should be pure).

val pre_compose_right_to_left : ('b, 'c) t -> ('a -> 'b) -> ('a, 'c) t

Reversed version of Preface_specs.Arrow.OPERATION.pre_compose_left_to_right.

val post_compose_right_to_left : ('b -> 'c) -> ('a, 'b) t -> ('a, 'c) t

Reversed version of Preface_specs.Arrow.OPERATION.post_compose_left_to_right.

Infix operators

module Infix : Preface_specs.Arrow.INFIX with type ('a, 'b) t = ('a, 'b) t

type ('a, 'b) t

The type held by the Arrow.

type ('a, 'b) t

The type held by the Category.

val (%) : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

An alias of Category.CORE.compose (to be iso with Preface_core).

val (<%) : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

An alias of Category.CORE.compose (to be iso with Preface_core).

val (%>) : ('a, 'b) t -> ('b, 'c) t -> ('a, 'c) t

An alias of Category.OPERATION.compose_left_to_right (to be iso with Preface_core).

val (<<<) : ('b, 'c) t -> ('a, 'b) t -> ('a, 'c) t

An alias of Category.CORE.compose (to be iso with Haskell's approach). Even <<< looks like <% (it is an alias for the same function), they differ in their priorities. OCaml documentation of operators priorities

val (>>>) : ('a, 'b) t -> ('b, 'c) t -> ('a, 'c) t

An alias of Preface_specs__.Category.CORE.compose_left_to_right (to be iso with Haskell's approach). Even >>> looks like %> (it is an alias for the same function), they differ in their priorities. OCaml documentation of operators priorities

val (***) : ('a, 'b) t -> ('c, 'd) t -> ('a * 'c, 'b * 'd) t

Infix version of Preface_specs.Arrow.CORE.split.

val (&&&) : ('a, 'b) t -> ('a, 'c) t -> ('a, 'b * 'c) t

Infix version of Preface_specs.Arrow.OPERATION.fan_out.

val (^>>) : ('a -> 'b) -> ('b, 'c) t -> ('a, 'c) t

Infix version of Preface_specs.Arrow.OPERATION.pre_compose_left_to_right.

val (>>^) : ('a, 'b) t -> ('b -> 'c) -> ('a, 'c) t

Infix version of Preface_specs.Arrow.OPERATION.post_compose_left_to_right.

val (<<^) : ('b, 'c) t -> ('a -> 'b) -> ('a, 'c) t

Infix version of Preface_specs.Arrow.OPERATION.post_compose_right_to_left.

val (^<<) : ('b -> 'c) -> ('a, 'b) t -> ('a, 'c) t

Infix version of Preface_specs.Arrow.OPERATION.post_compose_right_to_left.