Module Preface_make__Comonad

Building a Preface_specs.Comonad

Using the minimal definition

Using extract and extend

Build a Preface_specs.COMONAD using Preface_specs.Comonad.WITH_EXTEND.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_extend : functor (Req : Preface_specs.Comonad.WITH_EXTEND) -> Preface_specs.COMONAD with type 'a t = 'a Req.t

Using extract, map and duplicate

Build a Preface_specs.COMONAD using Preface_specs.Comonad.WITH_MAP_AND_DUPLICATE.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_map_and_duplicate : functor (Req : Preface_specs.Comonad.WITH_MAP_AND_DUPLICATE) -> Preface_specs.COMONAD with type 'a t = 'a Req.t

Using extract and the cokleisli composition

Build a Preface_specs.COMONAD using Preface_specs.Comonad.WITH_COKLEISLI_COMPOSITION.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_cokleisli_composition : functor (Req : Preface_specs.Comonad.WITH_COKLEISLI_COMPOSITION) -> Preface_specs.COMONAD with type 'a t = 'a Req.t

Manual construction

Advanced way to build a Preface_specs.COMONAD, constructing and assembling a component-by-component of Preface_specs.COMONAD. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Comonad.CORE) -> functor (Operation : Preface_specs.Comonad.OPERATION with type 'a t = 'a Core.t) -> functor (Syntax : Preface_specs.Comonad.SYNTAX with type 'a t = 'a Core.t) -> functor (Infix : Preface_specs.Comonad.INFIX with type 'a t = 'a Core.t) -> Preface_specs.COMONAD with type 'a t = 'a Core.t

Building Core

module Core_via_map_and_duplicate : functor (Req : Preface_specs.Comonad.WITH_MAP_AND_DUPLICATE) -> Preface_specs.Comonad.CORE with type 'a t = 'a Req.t

module Core_via_extend : functor (Req : Preface_specs.Comonad.WITH_EXTEND) -> Preface_specs.Comonad.CORE with type 'a t = 'a Req.t

module Core_via_cokleisli_composition : functor (Req : Preface_specs.Comonad.WITH_COKLEISLI_COMPOSITION) -> Preface_specs.Comonad.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Comonad.CORE) -> Preface_specs.Comonad.OPERATION with type 'a t = 'a Core.t

Deriving Syntax

module Syntax : functor (Core : Preface_specs.Comonad.CORE) -> Preface_specs.Comonad.SYNTAX with type 'a t = 'a Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Comonad.CORE) -> functor (Operation : Preface_specs.Comonad.OPERATION with type 'a t = 'a Core.t) -> Preface_specs.Comonad.INFIX with type 'a t = 'a Core.t