Module Preface_make__Divisible

Building a Preface_specs.Divisible

Using the minimal definition

Using divide and conquer

Build a Preface_specs.DIVISBLE using Preface_specs.Divisible.WITH_DIVIDE_AND_CONQUER. contramap will be derived using divide and conquer.

Standard method, using the minimal definition of a Divisible to derive its full API.

module Via_divide_and_conquer : functor (Req : Preface_specs.Divisible.WITH_DIVIDE_AND_CONQUER) -> Preface_specs.DIVISIBLE with type 'a t = 'a Req.t

Using divide, conquer and contramap

Build a Preface_specs.DIVISBLE using Preface_specs.Divisible.WITH_CONTRAMAP_AND_DIVIDE_AND_CONQUER.

Standard method, using the minimal definition of a Divisible to derive its full API.

module Via_contramap_and_divide_and_conquer : functor (Req : Preface_specs.Divisible.WITH_CONTRAMAP_AND_DIVIDE_AND_CONQUER) -> Preface_specs.DIVISIBLE with type 'a t = 'a Req.t

Over Contravariant functor

Produces a Preface_specs.DIVISBLE from a Preface_specs.CONTRAVARIANT.

module Over_contravariant : functor (Contravariant : Preface_specs.Contravariant.CORE) -> functor (Req : Preface_specs.Divisible.WITH_DIVIDE_AND_CONQUER with type 'a t = 'a Contravariant.t) -> Preface_specs.DIVISIBLE with type 'a t = 'a Req.t

Manual construction

Advanced way to build an Preface_specs.DIVISBLE, constructing and assembling a component-by-component of Preface_specs.DIVISBLE. (In order to provide your own implementation for some features.)

Grouping of all components

module Via : functor (Core : Preface_specs.Divisible.CORE) -> functor (Operation : Preface_specs.Divisible.OPERATION with type 'a t = 'a Core.t) -> functor (Infix : Preface_specs.Divisible.INFIX with type 'a t = 'a Operation.t) -> Preface_specs.DIVISIBLE with type 'a t = 'a Infix.t

Building Core

module Core_via_divide_and_conquer : functor (Req : Preface_specs.Divisible.WITH_DIVIDE_AND_CONQUER) -> Preface_specs.Divisible.CORE with type 'a t = 'a Req.t

module Core_via_contramap_and_divide_and_conquer : functor (Req : Preface_specs.Divisible.WITH_CONTRAMAP_AND_DIVIDE_AND_CONQUER) -> Preface_specs.Divisible.CORE with type 'a t = 'a Req.t

Deriving Operation

module Operation : functor (Core : Preface_specs.Divisible.CORE) -> Preface_specs.Divisible.OPERATION with type 'a t = 'a Core.t

Deriving Infix

module Infix : functor (Core : Preface_specs.Divisible.CORE) -> functor (Operation : Preface_specs.Divisible.OPERATION with type 'a t = 'a Core.t) -> Preface_specs.Divisible.INFIX with type 'a t = 'a Operation.t