Module Preface_make__Free_applicative

Building a Preface_specs.Free_applicative

Using the minimal definition

Over a Functor

Build a Preface_specs.FREE_APPLICATIVE over a Preface_specs.FUNCTOR.

module Over_functor : functor (F : Preface_specs.Functor.CORE) -> Preface_specs.FREE_APPLICATIVE with type 'a f = 'a F.t

Over an Applicative

Build a Preface_specs.FREE_APPLICATIVE over an Preface_specs.APPLICATIVE with a Natural transformationwhich extracts the underlying value and a run function wich is the Natural transformation from the underlying applicative using the identity.

module Over_applicative : functor (A : Preface_specs.Applicative.CORE) -> sig ... end