Module Preface_make__Profunctor

Building a Preface_specs.Profunctor

Using the minimal definition

Using dimap

Build a Preface_specs.PROFUNCTOR using Preface_specs.Profunctor.WITH_DIMAP.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_dimap : functor (Req : Preface_specs.Profunctor.WITH_DIMAP) -> Preface_specs.PROFUNCTOR with type ('a, 'b) t = ('a, 'b) Req.t

Using contramap_fst and map_snd

Build a Preface_specs.PROFUNCTOR using Preface_specs.Profunctor.WITH_CONTRAMAP_FST_AND_MAP_SND.

Standard method, using the minimal definition of an alt to derive its full API.

module Via_contramap_fst_and_map_snd : functor (Req : Preface_specs.Profunctor.WITH_CONTRAMAP_FST_AND_MAP_SND) -> Preface_specs.PROFUNCTOR with type ('a, 'b) t = ('a, 'b) Req.t

Profunctor Algebra

Construction of Preface_specs.PROFUNCTOR by combining them.

Composition

Right-to-left composition of Preface_specs.PROFUNCTOR.

module Composition : functor (F : Preface_specs.PROFUNCTOR) -> functor (G : Preface_specs.PROFUNCTOR) -> sig ... end

From other abstraction

From a Monad

Produces a Preface_specs.PROFUNCTOR from a Preface_specs.MONAD.

module From_monad : functor (Monad : Preface_specs.Monad.CORE) -> Preface_specs.PROFUNCTOR with type ('a, 'b) t = 'a -> 'b Monad.t

From a Strong profunctor

Produces a Preface_specs.PROFUNCTOR from a Preface_specs.STRONG.

module From_strong : functor (Strong : Preface_specs.STRONG) -> Preface_specs.PROFUNCTOR with type ('a, 'b) t = ('a, 'b) Strong.t

From a Choice profunctor

Produces a Preface_specs.PROFUNCTOR from a Preface_specs.CHOICE.

module From_choice : functor (Choice : Preface_specs.CHOICE) -> Preface_specs.PROFUNCTOR with type ('a, 'b) t = ('a, 'b) Choice.t

From a Closed profunctor

Produces a Preface_specs.PROFUNCTOR from a Preface_specs.CLOSED.

module From_closed : functor (Closed : Preface_specs.CLOSED) -> Preface_specs.PROFUNCTOR with type ('a, 'b) t = ('a, 'b) Closed.t

Manual construction

Advanced way to build a Preface_specs.PROFUNCTOR, constructing and assembling a component-by-component of Preface_specs.PROFUNCTOR. (In order to provide your own implementation for some features.)

Building Core

module Core_via_dimap : functor (Req : Preface_specs.Profunctor.WITH_DIMAP) -> Preface_specs.Profunctor.CORE with type ('a, 'b) t = ('a, 'b) Req.t

module Core_via_contramap_fst_and_map_snd : functor (Req : Preface_specs.Profunctor.WITH_CONTRAMAP_FST_AND_MAP_SND) -> Preface_specs.Profunctor.CORE with type ('a, 'b) t = ('a, 'b) Req.t