Parameter Suite.L

include Preface_specs.Join_semilattice.API with type t = R.t

type t = R.t

The type held by the Join_semillatice.

Functions

Infix operators

include Preface_specs.Meet_semilattice.API with type t := t

Functions

Infix operators

Functions

include Preface_specs.Join_semilattice.WITH_JOIN with type t := t

val join : t -> t -> t

join x y is the least upper bound of {x,y}.

include Preface_specs.Meet_semilattice.WITH_MEET with type t := t

val meet : t -> t -> t

meet x y is the greatest lower bound of {x, y}.

Infix operators

module Infix : Preface_specs.Lattice.INFIX with type t := t

val (||) : t -> t -> t

Infix version of join.

val (&&) : t -> t -> t

Infix version of meet.