Module Preface_qcheck.Model

Describes all the elements required to build a test suite.

Suit description

module type SUITE = sig ... end

Sampling Data

module type T0 = sig ... end

Types of arity 0

Describes the set of prerequisites for building generators on arity-free types (notably the monoid hierarchy).

module type COVARIANT_0 = T0

Types of arity 1

module type COVARIANT_1 = sig ... end

module type CONTRAVARIANT_1 = sig ... end

Types of arity 2

module type COVARIANT_2 = sig ... end

module type PROFUNCTORIAL = sig ... end