Module Preface_specs.Bounded_join_semilattice

A Join_semilattice capture the notion of a set with least upper bound with a bottom . A bounded join semilattice is an idempotent commutative semigroup.

Laws

To have a predictable behaviour, the instance of Bounded_join_semilattice must obey some laws.

join a (join b c) = join (join a b) c

join a b = join b a

join a a = a

join a bottom = a

Minimal definition

module type WITH_BOTTOM = sig ... end

module type WITH_JOIN_AND_BOTTOM = sig ... end

Structure anatomy

module type CORE = WITH_JOIN_AND_BOTTOM

Basis operations

module type INFIX = sig ... end

Infix operators.

Complete API

module type API = sig ... end

The complete interface of a Bounded_join_semillatice.