Module type Bounded_lattice.WITH_BOUNDED_JOIN_LATTICE_AND_BOUNDED_MEET_LATTICE

include Bounded_join_semilattice.WITH_JOIN_AND_BOTTOM

include Bounded_join_semilattice.WITH_BOTTOM

type t

The type held by the Bounded_join_semilattice

val bottom : t

The bottom value represent the least element of the join semilattice

include Join_semilattice.WITH_JOIN with type t := t

val join : t -> t -> t

join x y is the least upper bound of {x,y}.

include Bounded_meet_semilattice.WITH_MEET_AND_TOP with type t := t

include Bounded_meet_semilattice.WITH_TOP with type t := t

val top : t

the top value represent the greatest element of the meet semilattice

include Meet_semilattice.WITH_MEET with type t := t

val meet : t -> t -> t

meet x y is the greatest lower bound of {x, y}.