Module type Bounded_meet_semilattice.WITH_MEET_AND_TOP

include WITH_TOP

type t

The type held by the Bounded_meet_semilattice

val top : t

the top value represent the greatest element of the meet semilattice

include Meet_semilattice.WITH_MEET with type t := t

val meet : t -> t -> t

meet x y is the greatest lower bound of {x, y}.