Module Preface_specs.Env

An Env comonad parametrized over an inner Comonad and an env (which is an arbitrary type). Env is a comonad transformer. Env is Coreader (the dual of Reader).

module type CORE = sig ... end

Operation of Env comonad parametrized over an inner comonad and env.

Complete API

module type API = sig ... end

The complete interface of a Env comonad which introduces the Comonad API into the Env API.