Module Preface_specs.Foldable

A Foldable is a data structure which can be fold. In other word, reduced to a summary value one element at a time

Minimal definition

module type WITH_FOLD_MAP = sig ... end

Minimal definition using fold_map'.

module type WITH_FOLD_RIGHT = sig ... end

Minimal definition using fold_right.

Structure anatomy

module type CORE = sig ... end

Basis operation.

module type OPERATION = sig ... end

Additional operations.

Complete API

module type API = sig ... end

The complete interface of a Foldable.

Additional references

Haskell's wiki of Foldable and Traversable

Haskell's documentation of Foldable