Module type Freer_selective.API

The complete interface of a Freer selective.

type 'a f

The parametric type (which, unlike a Preface_specs.Free_selective don't need to be a Preface_specs.Functor).

type _ t =

| Pure : 'a -> 'a t

| Select : (('b -> 'a, 'a) Stdlib.Either.t t * 'b f) -> 'a t

The type held by the Freer selective.

val promote : 'a f -> 'a t

Promote a value from the Preface_specs.Functor into the Freer selective.

module To_selective (Selective : Selective.API) : TO_SELECTIVE with type 'a t := 'a t and type 'a f := 'a f and type 'a selective := 'a Selective.t

The natural transformation from a Freer selective to an other Preface_specs.Selective.

module To_monoid (Monoid : Monoid.CORE) : TO_MONOID with type 'a t := 'a t and type 'a f := 'a f and type monoid := Monoid.t

The natural transformation from a Freer selective to a Preface_specs.Monoid.

Selective API

A Freer selective is also a Preface_specs.Selective.

Type

Type

Functions

val select : ('a, 'b) Stdlib.Either.t t -> ('a -> 'b) t -> 'b t

select e f apply f if e is Left. It allow to skip effect using Right.

val branch : ('a, 'b) Stdlib.Either.t t -> ('a -> 'c) t -> ('b -> 'c) t -> 'c t

branch is like select. It chooses between two effects.

include Indexed_applicative.WITH_PURE with type ('a, 'index) t := 'a t

include Indexed_apply.WITH_APPLY with type ('a, 'index) t := 'a t

val apply : ('a -> 'b) t -> 'a t -> 'b t

May apply a function wrapped into an t to a value also wrapped into an t.

include Indexed_applicative.WITH_PURE with type ('a, 'index) t := 'a t

include Indexed_apply.WITH_MAP_AND_PRODUCT with type ('a, 'index) t := 'a t

include Indexed_functor.WITH_MAP with type ('a, 'index) t := 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over t from 'a to 'b.

include Indexed_apply.WITH_PRODUCT with type ('a, 'index) t := 'a t

val product : 'a t -> 'b t -> ('a * 'b) t

Monoidal product between two t.

include Indexed_applicative.WITH_PURE with type ('a, 'index) t := 'a t

val pure : 'a -> 'a t

Lift a value from 'a into a new t.

include Indexed_apply.WITH_LIFT2 with type ('a, 'index) t := 'a t

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over t from 'a to 'b.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val replace : 'a -> 'b t -> 'a t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : 'a t -> unit t

Create a new t, replacing all values in the given functor by unit.

val if_ : bool t -> 'a t -> 'a t -> 'a t

Same of branch but using a Boolean as disjunction.

val bind_bool : bool t -> (bool -> 'a t) -> 'a t

Monad bind specialized for Boolean.

val when_ : bool t -> unit t -> unit t

Conditionally perform an effect.

val exists : ('a -> bool t) -> 'a list -> bool t

Selective version of List.exists.

val for_all : ('a -> bool t) -> 'a list -> bool t

Selective version of List.for_all.

val or_ : bool t -> bool t -> bool t

Or combinator.

val and_ : bool t -> bool t -> bool t

And combinator.

val while_ : bool t -> unit t

Keep checking an effectful condition while it holds.

Infix operators

module Infix : Indexed_selective.INFIX with type ('a, 'index) t := 'a t

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (<**>) : 'a t -> ('a -> 'b) t -> 'b t

Flipped Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (*>) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (<*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of OPERATION.replace.

val (<*?) : ('a, 'b) Stdlib.Either.t t -> ('a -> 'b) t -> 'b t

Infix version of CORE.select.

val (<||>) : bool t -> bool t -> bool t

Infix version of CORE.or_.

val (<&&>) : bool t -> bool t -> bool t

Infix version of CORE.and_.

Syntax

module Syntax : Indexed_selective.SYNTAX with type ('a, 'index) t := 'a t

include Indexed_apply.SYNTAX with type ('a, 'index) t := 'a t

val let+ : 'a t -> ('a -> 'b) -> 'b t

let operator for mapping.

val and+ : 'a t -> 'b t -> ('a * 'b) t

and operator for the monoidal product.