Module type Indexed_comonad.API

The complete interface of an Indexed Comonad.

Type

type ('a, 'index) t

The type held by the Indexed Comonad.

Functions

val duplicate : ('a, 'index) t -> (('a, 'index) t, 'index) t

Dual of join.

val map : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val extend : (('a, 'index) t -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Dual of bind.

val extract : ('a, 'index) t -> 'a

Extract a 'a from t. Dual of return.

val compose_left_to_right : (('a, 'index) t -> 'b) -> (('b, 'index) t -> 'c) -> ('a, 'index) t -> 'c

Composing monadic functions using Co-Kleisli Arrow (from left to right).

val lift : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Mapping over t from 'a to 'b.

val lift2 : ('a -> 'b -> 'c) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift3 : ('a -> 'b -> 'c -> 'd) -> ('a, 'index) t -> ('b, 'index) t -> ('c, 'index) t -> ('d, 'index) t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val compose_right_to_left : (('b, 'index) t -> 'c) -> (('a, 'index) t -> 'b) -> ('a, 'index) t -> 'c

Composing co-monadic functions using Co-Kleisli Arrow (from right to left).

val replace : 'a -> ('b, 'index) t -> ('a, 'index) t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : ('a, 'index) t -> (unit, 'index) t

Create a new t, replacing all values in the given functor by unit.

Infix operators

module Infix : INFIX with type ('a, 'index) t := ('a, 'index) t

val (=>>) : ('a, 'index) t -> (('a, 'index) t -> 'b) -> ('b, 'index) t

Infix flipped version of CORE.extend.

val (<<=) : (('a, 'index) t -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Infix version of CORE.extend.

val (=>=) : (('a, 'index) t -> 'b) -> (('b, 'index) t -> 'c) -> ('a, 'index) t -> 'c

Infix version of CORE.compose_left_to_right.

val (=<=) : (('b, 'index) t -> 'c) -> (('a, 'index) t -> 'b) -> ('a, 'index) t -> 'c

Infix version of OPERATION.compose_right_to_left.

val (<@@>) : ('a, 'index) t -> ('a -> 'b, 'index) t -> ('b, 'index) t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (<@>) : ('a -> 'b, 'index) t -> ('a, 'index) t -> ('b, 'index) t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (@>) : (unit, 'index) t -> ('b, 'index) t -> ('b, 'index) t

Discard the value of the first argument.

val (<@) : ('a, 'index) t -> (unit, 'index) t -> ('a, 'index) t

Discard the value of the second argument.

val (<$>) : ('a -> 'b) -> ('a, 'index) t -> ('b, 'index) t

Infix version of CORE.map.

val (<&>) : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

Flipped and infix version of CORE.map.

val (<$) : 'a -> ('b, 'index) t -> ('a, 'index) t

Infix version of OPERATION.replace.

val ($>) : ('a, 'index) t -> 'b -> ('b, 'index) t

Flipped and infix version of OPERATION.replace.

Syntax

module Syntax : SYNTAX with type ('a, 'index) t := ('a, 'index) t

val let@ : ('a, 'index) t -> (('a, 'index) t -> 'b) -> ('b, 'index) t

Syntactic shortcuts for version of CORE.extend:

let@ x = e in f is equals to extend f e.

include Indexed_functor.SYNTAX with type ('a, 'index) t := ('a, 'index) t

val let+ : ('a, 'index) t -> ('a -> 'b) -> ('b, 'index) t

let operator for mapping.