Module Preface_specs.Indexed_functor

An Indexed Functor represents a type that can be mapped over (with an index, so an additional type parameter that can be useful for tracking information using row polymorphism). So we can go from ('a, 'index) t to ('b, 'index) t using a function from 'a to 'b. Mapping preserve the structure of the input.

Minimal definition

Minimal definition

module type WITH_MAP = sig ... end

The minimum definition of an Indexed Functor. It is by using the combinators of this module that the other combinators will be derived.

Structure anatomy

module type CORE = WITH_MAP

Basis operations.

module type OPERATION = sig ... end

Additional operations.

module type INFIX = sig ... end

Infix operators.

module type SYNTAX = sig ... end

Syntax operators.

Complete API

module type API = sig ... end

The complete interface of a Functor.