Module type Monad_plus.WITH_MAP_AND_JOIN

Minimal definition using neutral, combine, return, map and join.

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val join : 'a t t -> 'a t

join remove one level of monadic structure, projecting its bound argument into the outer level.

type 'a t

A type 'a t held by the Alt.

val combine : 'a t -> 'a t -> 'a t

Combine two values of 'a t into one.

val neutral : 'a t

The neutral element of the Alternative.