Module Preface_specs.State

A State monad parametrized over an inner Monad and a state (which is an arbitrary type). State is a monad transformer.

module type CORE = sig ... end

Operation of State monad parametrized over an inner monad and state.

Complete API

module type API = sig ... end

The complete interface of a State monad which introduces the Monad API into the State API.