Module Preface_specs.Store

A Store comonad parametrized over an inner Comonad and a store (which is an arbitrary type). Store is a comonad transformer. Store is Costate (the dual of State).

module type CORE = sig ... end

Operation of Store comonad parametrized over an inner comonad and store.

Complete API

module type API = sig ... end

The complete interface of a Store comonad which introduces the Comonad API into the Store API.