Module API.Experiment

Applies a function which lift the store to a functor-value and use the accessor to read the resulting focus.

Parameters

module F : Functor.API

Signature

val run : (store -> store F.t) -> 'a t -> 'a F.t

Perform the accessor for reading the resulting focus.