Module Preface_specs.Traversable

A Traversable is a data structure that can be traversed from left to right, performing an action on each element.

A common usage of Traversable is to turn any Traversable of Applicative into a Applicative of Traversable.

For example, going to 'a option list to 'a list option.

Minimal definition

module type WITH_TRAVERSE = sig ... end

Minimal definition using traverse over a 'a iter.

Structure anatomy

module type CORE = WITH_TRAVERSE

Basis operations.

module type OPERATION = sig ... end

Additional operations.

Complete API

module type API = sig ... end

The complete interface of a Traversable

module type API_OVER_MONAD = sig ... end

The complete interface of a Traversable over a Monad

module type API_OVER_APPLICATIVE = sig ... end

The complete interface of a Traversable over an Applicative

Additional references

Applicative Programming with Effects

The Essence of the Iterator Pattern

An Investigation of the Laws of Traversals