Module type Traversable.API

The complete interface of a Traversable

Types

type 'a t

The type held by the Traversable.

type 'a iter

The iterable type held by the Traversable.

Functions

val traverse : ('a -> 'b t) -> 'a iter -> 'b iter t

Map each element of a structure to an action, evaluate these actions from left to right, and collect the results. *

val sequence : 'a t iter -> 'a iter t

Evaluate each action in the structure from left to right, and collect the results