Module type Traversable.WITH_TRAVERSE

Minimal definition using traverse over a 'a iter.

type 'a t

The type held by the Traversable.

type 'a iter

The iterable type held by the Traversable.

val traverse : ('a -> 'b t) -> 'a iter -> 'b iter t

Map each element of a structure to an action, evaluate these actions from left to right, and collect the results. *