Module type Arrow_choice.WITH_LEFT_AND_CHOOSE

Minimal definition using choose and left without Arrow.

type ('a, 'b) t

The type held by the Arrow_choice.

val left : ('a, 'b) t -> (('a, 'c) Preface_core.Shims.Either.t, ('b, 'c) Preface_core.Shims.Either.t) t

Feed marked inputs through the argument arrow, passing the rest through unchanged to the output.

type ('a, 'b) t

The type held by the Arrow_choice.

val choose : ('a, 'b) t -> ('c, 'd) t -> (('a, 'c) Preface_core.Shims.Either.t, ('b, 'd) Preface_core.Shims.Either.t) t

Split the input between the two argument arrows, retagging and merging their outputs.