Module Preface_specs__.Arrow_plus

Laws

To have a predictable behaviour, the instance of Arrow_plus must obey some laws.

All Arrow_alt laws

Structure anatomy

module type WITH_COMBINE_AND_NEUTRAL = sig ... end

Exposes the neutral value and combine function, mandatory for each requirement.

module type WITH_ARROW_AND_FST = sig ... end

RMinimal definition using arrow, fst, neutral and combine.

module type WITH_ARROW_AND_SPLIT = sig ... end

Minimal definition using arrow, split, neutral and combine.

Structure anatomy

module type CORE = sig ... end

Basis operations.

module type OPERATION = sig ... end

Additional operations.

module type ALIAS = Preface_specs.Arrow.ALIAS

Aliases of some operations functions.

module type INFIX = Preface_specs.Arrow_alt.INFIX

Infix operators.

Complete API

module type API = sig ... end

The complete interface of an Arrow_plus.

Additional references

Generalising Monads to Arrows

Arrows: A General Interface to Computation

Haskell's documentation of Arrow