Module type Free_applicative.CORE

The Free applicative API without the Preface_specs.Applicative API.

type 'a f

The type held by the Preface_specs.Functor.

type _ t =

| Pure : 'a -> 'a t

| Apply : ('a -> 'b) t * 'a f -> 'b t

The type held by the Free applicative.

val promote : 'a f -> 'a t

Promote a value from the Preface_specs.Functor into the Free applicative.

module To_applicative : functor (Applicative : Preface_specs.Applicative.CORE) -> TO_APPLICATIVE with type 'a t := 'a t and type 'a f := 'a f and type 'a applicative := 'a Applicative.t

The natural transformation from a Free applicative to an other Preface_specs.Applicative.

module To_monoid : functor (Monoid : Preface_specs.Monoid.CORE) -> TO_MONOID with type 'a t := 'a t and type 'a f := 'a f and type monoid := Monoid.t

The natural transformation from a Free applicative to a Preface_specs.Monoid.