Module Preface_specs__.Free_monad

Note about complexity

Although free constructs are elegant, they introduce an execution cost due to the recursive nature of defining the type of a Free Monad. There are cheaper encodings like Preface_specs.Freer_monad, which removes the requirement for the Free monad parameter to be a Preface_specs.Functor and therefore does not increase its complexity.

Structure anatomy

module type CORE = sig ... end

The Free Monad API without the Preface_specs.Monad API.

Complete API

module type API = sig ... end

The complete interface of a Free monad.