Module Preface_specs__.Semigroup

Laws

To ensure that the derived combiners work properly, the combine function must comply with the following laws:

combine (combine a b) c = combine a (combine b c)

Minimal definition

module type WITH_COMBINE = sig ... end

The minimum definition of a Semigroup. It is by using the combinators of this module that the other combinators will be derived.

Structure anatomy

module type CORE = WITH_COMBINE

Basis operations.

module type OPERATION = sig ... end

Additional operations.

module type INFIX = sig ... end

Infix operators.

Complete API

module type API = sig ... end

The complete interface of a Semigroup.

Additional references

Haskell's documentation of a Semigroup