Module Preface_specs__.State

module type CORE = sig ... end

Operation of State monad parametrized over an inner monad and state.

Complete API

module type API = sig ... end

The complete interface of a State monad wich introduces the Monad API into the State API.