Module Preface_specs__.Store

module type CORE = sig ... end

Operation of Store comonad parametrized over an inner comonad and store.

Complete API

module type API = sig ... end

The complete interface of a Store comonad wich introduces the Comonad API into the Store API.