Module type Writer.API

The complete interface of a Writer monad wich introduces the Monad API into the Writer API.

type tape

The encapsulated tape.

type 'a monad

The inner monad.

type 'a t = ('a * tape) monad

The type held by the writer monad.

type 'a monad

The inner monad.

type 'a t

The type held by the monad transformer.

val upper : 'a monad -> 'a t

promote monad into the transformation.

val writer : ('a * tape) -> 'a t

Construct a writer computation from a (result, output) pair.

val run : 'a t -> ('a * tape) monad

Unwrap the writer computation.

val exec : 'a t -> tape monad

Extract the output from the writer computation.

val tell : tape -> unit t

tell helps to enrich the output. This is done thanks to the semigroup combine operation.

val listen : 'a t -> ('a * tape) t

listen executes the effect and return both the result and the corresponding output within the effect.

val listens : (tape -> 'b) -> 'a t -> ('a * 'b) t

Performs the action and adds the result of applying the function to the output to the value of the computation.

val pass : ('a * (tape -> tape)) t -> 'a t

pass executes the effect and apply the function to the corresponding output.

val censor : (tape -> tape) -> 'a t -> 'a t

censor executes the effects, apply the function to the corresponding output and returns an effects with the value unchanged.

Monad

module Monad : Preface_specs.Monad.API

Type

type 'a t

The type held by the Monad.

Functions

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val bind : ('a -> 'b t) -> 'a t -> 'b t

bind f m passes the result of computation m to function f.

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val join : 'a t t -> 'a t

join remove one level of monadic structure, projecting its bound argument into the outer level.

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val compose_left_to_right : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from left to right).

type 'a t

The type held by the Monad.

val compose_right_to_left : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from right to left).

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Mapping over from 'a and 'b to 'c over 'a t and 'b t to 'c t.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Mapping over from 'a and 'b and 'c to 'd over 'a t and 'b t and 'c t to 'd t.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

Infix operators

module Infix = Monad.Infix

type 'a t

The type held by the Monad.

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Monad.CORE.map.

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

Infix flipped version of Monad.CORE.map.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

Infix flipped version of Monad.CORE.bind.

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

Infix version of Monad.CORE.bind.

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Infix version of Monad.CORE.compose_left_to_right.

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Infix version of Monad.OPERATION.compose_right_to_left.

val (>>) : unit t -> 'b t -> 'b t

Sequentially compose two actions, discarding any value produced by the first.

val (<<) : 'a t -> unit t -> 'a t

Sequentially compose two actions, discarding any value produced by the second.

type 'a t

The type held by the Functor.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Preface_specs.Functor.CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Preface_specs.Functor.CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Preface_specs.Functor.OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Preface_specs.Functor.OPERATION.replace.

Syntax

module Syntax = Monad.Syntax

type 'a t

The type held by the Monad.

val let* : 'a t -> ('a -> 'b t) -> 'b t

Syntaxic shortcuts for flipped version of Monad.CORE.bind:

let* x = e in f is equals to bind (fun x -> f) e.

val let+ : 'a t -> ('a -> 'b) -> 'b t

Syntaxic shortcuts for flipped version of Monad.CORE.map:

let+ x = e in f is equals to map (fun x -> f) e.