Module Preface_specs__Arrow_alt

An Arrow with a combine function. So Arrow_alt is also an Arrow. Arrow_alt is a kind of Alt in the arrow hierarchy.

Laws

To have a predictable behaviour, the instance of Arrow_alt must obey some laws.

All Arrow laws

combine (combine a b) c = combine a (combine b c)

Minimal definition

module type WITH_COMBINE = sig ... end

Exposes the combine function, mandatory for each requirement.

module type WITH_ARROW_AND_FST = sig ... end

Minimal definition using arrow, fst and combine.

module type WITH_ARROW_AND_SPLIT = sig ... end

Minimal definition using arrow, split and combine.

Structure anatomy

module type CORE = sig ... end

Basis operations.

module type OPERATION = sig ... end

Additional operations.

module type ALIAS = Preface_specs.Arrow.ALIAS

Aliases of some operations functions.

module type INFIX = sig ... end

Infix operators.

Complete API

module type API = sig ... end

The complete interface of an Arrow_alt.

Additional references

Generalising Monads to Arrows

Arrows: A General Interface to Computation

Haskell's documentation of Arrow