Module Preface_specs__Divisible

Divisible is a "Contravariant Applicative Functor", in other word, Divisible is the dual of an Applicative

Minimal definition

module type WITH_DIVIDE_AND_CONQUER = sig ... end

Exposes the divide and conquer functions, mandatory for each requirement.

module type WITH_CONTRAMAP_AND_DIVIDE_AND_CONQUER = sig ... end

Structure anatomy

module type CORE = WITH_CONTRAMAP_AND_DIVIDE_AND_CONQUER

Basis operations.

module type OPERATION = sig ... end

Additional operations.

module type INFIX = sig ... end

Infix operators.

Complete API

module type API = sig ... end

The complete interface of a Divisible.

Additional references

https://typeclasses.com/contravariance