Module type Preface_specs__Foldable.WITH_FOLD_MAP

Minimal definition using fold_map'.

fold_map' neutral combine f x use explicit monoidal combinators passing in order to deal with polymorphsim.

type 'a t

The type held by Foldable.

val fold_map' : 'a -> ('a -> 'a -> 'a) -> ('b -> 'a) -> 'b t -> 'a

Map each element of the Foldable to a Monoid an combine the result.