Module Preface_specs__Free_selective

A Free selective allows you to build a rigid Preface_specs.Selective from a given Preface_specs.Functor.

Note about complexity

Although free constructs are elegant, they introduce an execution cost due to the recursive nature of defining the type of a Free Selective. There are cheaper encodings but they are not, for the moment, available in Preface.

Structure anatomy

module type TO_SELECTIVE = sig ... end

The natural transformation for Free Selective to Selective.

module type TO_MONOID = sig ... end

The natural transformation for Free Selective to Monoid.

module type CORE = sig ... end

The Free selective API without the Preface_specs.Selective API.

Complete API

module type API = sig ... end

The complete interface of a Free selective.

Additional references

Haskell's documentation of a Selective Application Functor

Selective Applicative Functors