Module type Preface_specs__Monad_plus.API

The complete interface of a Monad_plus.

Type

type 'a t

The type held by the Monad_plus.

Functions

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val bind : ('a -> 'b t) -> 'a t -> 'b t

bind f m passes the result of computation m to function f.

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val join : 'a t t -> 'a t

join remove one level of monadic structure, projecting its bound argument into the outer level.

type 'a t

The type held by the Monad.

val return : 'a -> 'a t

Create a new 'a t.

val compose_left_to_right : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from left to right).

type 'a t

A type 'a t held by the Alt.

val combine : 'a t -> 'a t -> 'a t

Combine two values of 'a t into one.

val neutral : 'a t

The neutral element of the Alternative.

type 'a t

The type held by the Monad.

val compose_right_to_left : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from right to left).

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Mapping over from 'a and 'b to 'c over 'a t and 'b t to 'c t.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Mapping over from 'a and 'b and 'c to 'd over 'a t and 'b t and 'c t to 'd t.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

type 'a t

A type 'a t held by the Alt.

val times : int -> 'a t -> 'a t option

times n x apply combine on x n times. If n is lower than 1 the function will returns None.

val reduce_nel : 'a t Preface_core.Nonempty_list.t -> 'a t

Reduce a Nonempty_list.t using combine.

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

val reduce : 'a t list -> 'a t

Reduce a List.t using combine.

val filter : ('a -> bool) -> 'a t -> 'a t

Filtering over Monad_plus.

Infix operators

module Infix : INFIX with type 'a t := 'a t

type 'a t

The type held by the Monad.

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Monad.CORE.map.

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

Infix flipped version of Monad.CORE.map.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

Infix flipped version of Monad.CORE.bind.

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

Infix version of Monad.CORE.bind.

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Infix version of Monad.CORE.compose_left_to_right.

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Infix version of Monad.OPERATION.compose_right_to_left.

val (>>) : unit t -> 'b t -> 'b t

Sequentially compose two actions, discarding any value produced by the first.

val (<<) : 'a t -> unit t -> 'a t

Sequentially compose two actions, discarding any value produced by the second.

type 'a t

The type held by the Functor.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Preface_specs.Functor.CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Preface_specs.Functor.CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Preface_specs.Functor.OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Preface_specs.Functor.OPERATION.replace.

val (<|>) : 'a t -> 'a t -> 'a t

Infix version of CORE.combine.

Syntax

module Syntax : SYNTAX with type 'a t := 'a t

type 'a t

The type held by the Monad.

val let* : 'a t -> ('a -> 'b t) -> 'b t

Syntaxic shortcuts for flipped version of Monad.CORE.bind:

let* x = e in f is equals to bind (fun x -> f) e.

val let+ : 'a t -> ('a -> 'b) -> 'b t

Syntaxic shortcuts for flipped version of Monad.CORE.map:

let+ x = e in f is equals to map (fun x -> f) e.