Module Preface_specs__Reader

A Reader monad parametrized over an inner Monad and an environment. Reader is a monad transformer.

Structure anatomy

module type CORE = sig ... end

Operation of Reader monad parametrized over an inner monad and an environment.

Complete API

module type API = sig ... end

The complete interface of a Reader monad wich introduces the Monad API into the Reader API.