Module Preface_specs__Traced

A Traced comonad parametrized over an inner Comonad and a tape (which is an arbitrary type). Traced is a comonad transformer. Traced is Cowriter (the dual of Writer).

module type CORE = sig ... end

Operation of Traced comonad parametrized over an inner comonad and tape.

Complete API

module type API = sig ... end

The complete interface of a Traced comonad wich introduces the Comonad API into the Traced API.