Module Preface_stdlib.Env

A specialized version of an Env comonad with Identity as the inner comonad. Since Preface.Make.Env is a Transformer, this module exposes the classical Env comonad (or Coreader).

Specialization

An Env Comonad is set up by a type.

module Over (Env : Preface_specs.Types.T0) : sig ... end