Parameter Traversable.A

Type

type 'a t

The type held by the Applicative.

Type

Functions

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_apply.WITH_APPLY with type ('a, 'index) t := 'a t

val apply : ('a -> 'b) t -> 'a t -> 'b t

May apply a function wrapped into an t to a value also wrapped into an t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_apply.WITH_MAP_AND_PRODUCT with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_functor.WITH_MAP with type ('a, 'index) t := 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over t from 'a to 'b.

include Preface_specs.Indexed_apply.WITH_PRODUCT with type ('a, 'index) t := 'a t

val product : 'a t -> 'b t -> ('a * 'b) t

Monoidal product between two t.

include Preface_specs.Indexed_applicative.WITH_PURE with type ('a, 'index) t := 'a t

val pure : 'a -> 'a t

Lift a value from 'a into a new t.

include Preface_specs.Indexed_apply.WITH_LIFT2 with type ('a, 'index) t := 'a t

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over t from 'a to 'b.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val replace : 'a -> 'b t -> 'a t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : 'a t -> unit t

Create a new t, replacing all values in the given functor by unit.

Infix operators

module Infix : Preface_specs.Indexed_applicative.INFIX with type ('a, 'index) t := 'a t

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (<**>) : 'a t -> ('a -> 'b) t -> 'b t

Flipped Applicative functor of ('a -> 'b) t over ('a, 'index) t to ('b, 'index) t.

val (*>) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (<*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of OPERATION.replace.

Syntax

module Syntax : Preface_specs.Indexed_applicative.SYNTAX with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_apply.SYNTAX with type ('a, 'index) t := 'a t

val let+ : 'a t -> ('a -> 'b) -> 'b t

let operator for mapping.

val and+ : 'a t -> 'b t -> ('a * 'b) t

and operator for the monoidal product.