Module Preface_stdlib.Reader

A specialized version of a Reader monad with Identity as the inner monad. Since Preface.Make.Reader is a Transformer, this module exposes the classical Reader monad.

Specialization giving an Environment

A Reader Monad is set up by the environment which be requested.

module Over (Env : Preface_specs.Types.T0) : sig ... end