Module Result.Mono

Parameters

module T : Preface_specs.Types.T0

Signature

Delayed implementation

By setting the error type of Result.t it is possible to get implementations for abstractions on constructors of type with an arity of 1.

Functor

module Functor : Preface_specs.FUNCTOR with type 'a t = ('a, T.t) Bifunctor.t

Alt

module Alt : Preface_specs.ALT with type 'a t = ('a, T.t) Bifunctor.t

Invariant

module Invariant : Preface_specs.INVARIANT with type 'a t = ('a, T.t) Bifunctor.t

Applicative

Result.t implements Preface_specs.APPLICATIVE and introduces an interface to define Preface_specs.TRAVERSABLE using Result as an iterable structure.

module Applicative : Preface_specs.Traversable.API_OVER_APPLICATIVE with type 'a t = ('a, T.t) Bifunctor.t

Selective

module Selective : Preface_specs.SELECTIVE with type 'a t = ('a, T.t) Bifunctor.t

Monad

Result.t implements Preface_specs.MONAD and introduces an interface to define Preface_specs.TRAVERSABLE using Result as an iterable structure.

module Monad : Preface_specs.Traversable.API_OVER_MONAD with type 'a t = ('a, T.t) Bifunctor.t

Foldable

module Foldable : Preface_specs.FOLDABLE with type 'a t = ('a, T.t) Bifunctor.t