Module Preface_stdlib.State

A specialized version of a State monad with Identity as the inner monad. Since Preface.Make.State is a Transformer, this module exposes the classical State monad.

Specialization

A State Monad is set up by a type.

module Over (State : Preface_specs.Types.T0) : sig ... end