Module Preface_stdlib.Store

A specialized version of a Store comonad with Identity as the inner comonad. Since Preface.Make.Store is a Transformer, this module exposes the classical Store comonad (or Costate).

Specialization

A Store Comonad is set up by a type.

module Over : functor (Store : Preface_specs.Types.T0) -> sig ... end