Module Preface_stdlib.Traced

A specialized version of a Traced comonad with Identity as the inner comonad. Since Preface.Make.Traced is a Transformer, this module exposes the classical Traced comonad (or Cowriter).

Specialization

A Traced Comonad is set up by a monoid.

module Over : functor (Tape : Preface_specs.MONOID) -> sig ... end