Module Preface_stdlib.Writer

A specialized version of a Writer monad with Identity as the inner monad. Since Preface.Make.Writer is a Transformer, this module exposes the classical Writer monad.

Specialization giving a Monoid

A Writer Monad is set up by the monoidal tape which be written.

module Over (Tape : Preface_specs.MONOID) : sig ... end