Module Applicative.Traversable

Parameters

A : Preface_specs.Applicative.API

Signature

Types

type 'a t = 'a A.t

The type held by the Traversable.

type 'a iter = 'a t

The iterable type held by the Traversable.

Functions

type 'a t

The type held by the Traversable.

type 'a iter

The iterable type held by the Traversable.

val traverse : ('a -> 'b t) -> 'a iter -> 'b iter t

Map each element of a structure to an action, evaluate these actions from left to right, and collect the results. *

type 'a t

The type held by the Traversable.

type 'a iter

The iterable type held by the Traversable.

val sequence : 'a t iter -> 'a iter t

Evaluate each action in the structure from left to right, and collect the results