Module Preface_stdlib__Approximation

Build monoidal approximation.

Over approximation

The selective functor Over can be used for computing a list of all effects embedded in a computation, i.e. an "over-approximation" of the effects that will actually occur. This is achieved by keeping track of effects in both arguments of the select operator.

module Over : functor (M : Preface_specs.MONOID) -> sig ... end

Under approximation

The selective functor Under discards the second argument of select, and therefore computes an "under-approximation", i.e. a list of effects that are guaranteed to occur.

module Under : functor (M : Preface_specs.MONOID) -> sig ... end