Module Free_monad.Over_selective

Parameters

module F : Preface_specs.SELECTIVE

Signature

type 'a f = 'a F.t

The type held by the Preface_specs.Functor.

type 'a t =

| Return of 'a

| Bind of 'a t f

The type held by the Free monad.

val perform : 'a f -> 'a t

Create a new 'a t from a 'a f.

val run : ('a f -> 'a) -> 'a t -> 'a

Execute a given handler for given data

module To_monad (Monad : Preface_specs.Monad.CORE) : Preface_specs.Free_monad.TO_MONAD with type 'a t := 'a t and type 'a f := 'a f and type 'a monad := 'a Monad.t

Functor API

A Free monad is also a Preface_specs.Functor.

module Functor : Preface_specs.Functor.API with type 'a t = 'a t

Applicative API

A Free monad is also an Preface_specs.Applicative.

module Applicative : Preface_specs.Applicative.API with type 'a t = 'a t

Selective API

A Free monad is also an Preface_specs.Selective.

module Selective : Preface_specs.Selective.API with type 'a t = 'a t

Monad API

A Free monad is also (obviously) a Preface_specs.Monad.

module Monad : Preface_specs.Monad.API with type 'a t = 'a t

Type

Type

Functions

include Preface_specs.Indexed_monad.WITH_RETURN with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_bind.WITH_BIND with type ('a, 'index) t := 'a t

val bind : ('a -> 'b t) -> 'a t -> 'b t

bind f m passes the result of computation m to function f.

include Preface_specs.Indexed_monad.WITH_RETURN with type ('a, 'index) t := 'a t

include Preface_specs.Indexed_bind.WITH_MAP_AND_JOIN with type ('a, 'index) t := 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b.

val join : 'a t t -> 'a t

join remove one level of monadic structure, projecting its bound argument into the outer level.

include Preface_specs.Indexed_monad.WITH_RETURN with type ('a, 'index) t := 'a t

val return : 'a -> 'a t

Lift a value into a t.

include Preface_specs.Indexed_bind.WITH_KLEISLI_COMPOSITION with type ('a, 'index) t := 'a t

val compose_left_to_right : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from left to right).

val compose_right_to_left : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Composing monadic functions using Kleisli Arrow (from right to left).

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over t from 'a to 'b.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Lift a binary function that acts on arbitrary values into a function that acts t values.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Lift a ternary function that acts on arbitrary values into a function that acts t values.

val replace : 'a -> 'b t -> 'a t

Create a new t, replacing all values of the given functor by given a value of 'a.

val void : 'a t -> unit t

Create a new t, replacing all values in the given functor by unit.

Infix operators

module Infix : Preface_specs.Indexed_monad.INFIX with type ('a, 'index) t := 'a t

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

Infix version of CORE.map.

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

Infix flipped version of CORE.map.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

Infix flipped version of CORE.bind.

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

Infix version of CORE.bind.

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Infix version of CORE.compose_left_to_right.

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Infix version of OPERATION.compose_right_to_left.

val (>>) : unit t -> 'b t -> 'b t

Sequentially compose two actions, discarding any value produced by the first.

val (<<) : 'a t -> unit t -> 'a t

Sequentially compose two actions, discarding any value produced by the second.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of OPERATION.replace.

Syntax

module Syntax : Preface_specs.Indexed_monad.SYNTAX with type ('a, 'index) t := 'a t

val let* : 'a t -> ('a -> 'b t) -> 'b t

Syntactic shortcuts for flipped version of CORE.bind:

let* x = e in f is equals to bind (fun x -> f) e.

val let+ : 'a t -> ('a -> 'b) -> 'b t

let operator for mapping.