Module Store.Over_comonad

Parameters

C : Preface_specs.COMONAD

Store : Preface_specs.Types.T0

Signature

type store

The encapsulated store.

type 'a comonad

The inner comonad.

type 'a t = (store -> 'a) comonad * store

The type held by the store comonad.

type 'a comonad

The inner comonad.

type 'a t

The type held by the comonad transformer.

val lower : 'a t -> 'a comonad

get the underlying comonad.

val run : 'a t -> (store -> 'a) comonad * store

Unwrap the store computation.

val pos : 'a t -> store

Read the stored value.

val peek : store -> 'a t -> 'a

Get the current focus for a different stored value (absolute getter).

val peeks : (store -> store) -> 'a t -> 'a

Get the current focus for a different modified stored value (absolute setter).

val seek : store -> 'a t -> 'a t

Move the current focus (absolute setter).

val seeks : (store -> store) -> 'a t -> 'a t

Modify (set and apply a function) the focus (relative setter).

module Experiment : functor (F : Preface_specs.Functor.API) -> sig ... end

Applies a function which lift the store to a functor-value and use the accessor to read the resulting focus.

Comonad

module Comonad : Preface_specs.Comonad.API

Type

type 'a t

The type held by the Comonad.

Functions

type 'a t

The type held by the Comonad.

val extract : 'a t -> 'a

Extract a 'a from 'a t. Dual of return.

val duplicate : 'a t -> 'a t t

Dual of join.

val map : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

type 'a t

The type held by the Comonad.

val extract : 'a t -> 'a

Extract a 'a from 'a t. Dual of return.

val extend : ('a t -> 'b) -> 'a t -> 'b t

Dual of bind.

type 'a t

The type held by the Comonad.

val extract : 'a t -> 'a

Extract a 'a from 'a t. Dual of return.

val compose_left_to_right : ('a t -> 'b) -> ('b t -> 'c) -> 'a t -> 'c

Composing monadic functions using Co-Kleisli Arrow (from left to right).

type 'a t

The type held by the Comonad.

val lift : ('a -> 'b) -> 'a t -> 'b t

Mapping over from 'a to 'b over 'a t to 'b t.

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Mapping over from 'a and 'b to 'c over 'a t and 'b t to 'c t.

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

Mapping over from 'a and 'b and 'c to 'd over 'a t and 'b t and 'c t to 'd t.

val compose_right_to_left : ('b t -> 'c) -> ('a t -> 'b) -> 'a t -> 'c

Composing comonadic functions using Co-Kleisli Arrow (from right to left).

type 'a t

The type held by the Functor.

val replace : 'a -> 'b t -> 'a t

Create a new 'a t, replacing all values in the 'b t by given a value of 'a.

val void : 'a t -> unit t

Create a new unit t, replacing all values in the 'a t by unit.

Infix operators

module Infix = Comonad.Infix

type 'a t

The type held by the Comonad.

val (=>>) : 'a t -> ('a t -> 'b) -> 'b t

Infix flipped version of Comonad.CORE.extend.

val (<<=) : ('a t -> 'b) -> 'a t -> 'b t

Infix version of Comonad.CORE.extend.

val (=>=) : ('a t -> 'b) -> ('b t -> 'c) -> 'a t -> 'c

Infix version of Comonad.CORE.compose_left_to_right.

val (=<=) : ('b t -> 'c) -> ('a t -> 'b) -> 'a t -> 'c

Infix version of Comonad.OPERATION.compose_right_to_left.

val (<@@>) : 'a t -> ('a -> 'b) t -> 'b t

Applicative functor of ('a -> 'b) t over 'a t to 'b t.

val (<@>) : ('a -> 'b) t -> 'a t -> 'b t

Applicative functor of ('a -> 'b) t over 'a t to 'b t.

val (@>) : unit t -> 'b t -> 'b t

Discard the value of the first argument.

val (<@) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

type 'a t

The type held by the Functor.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Preface_specs.Functor.CORE.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Preface_specs.Functor.CORE.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Preface_specs.Functor.OPERATION.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Preface_specs.Functor.OPERATION.replace.

Syntax

module Syntax = Comonad.Syntax

type 'a t

The type held by the Comonad.

val let@ : 'a t -> ('a t -> 'b) -> 'b t

Syntaxic shortcuts for version of Comonad.CORE.extend:

let@ x = e in f is equals to extend f e.

val let+ : 'a t -> ('a -> 'b) -> 'b t

Syntaxic shortcuts for version of Comonad.CORE.map