Module Preface_specs.Category

A Category is a Semigroupoid with an identity at each object.

Laws

To have a predictable behaviour, the instance of Category must obey some laws.

f % id = f

id % f = f

f % (g % h) = (f % g) % h

Minimal definition

module type WITH_ID = sig ... end

Minimal interface using id.

module type WITH_ID_AND_COMPOSE = sig ... end

Minimal interface using id and compose.

Structure anatomy

module type CORE = WITH_ID_AND_COMPOSE

Basis operations.

module type OPERATION = sig ... end

Additional operations.

module type INFIX = sig ... end

Infix operators.

Complete API

module type API = sig ... end

The complete interface of a Category.

Additional references

Haskell's documentation of Category